% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_china_from_provinces_report}
\alias{add_china_from_provinces_report}
\title{add_china_from_provinces_report}
\usage{
add_china_from_provinces_report(report)
}
\description{
For GCAM-China reporting, synthesize China rows from provincial rows for
additive variables when the China row is missing or empty.
}
\keyword{internal}
