% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_gcam_china_run_note}
\alias{add_gcam_china_run_note}
\title{add_gcam_china_run_note}
\usage{
add_gcam_china_run_note(note)
}
\arguments{
\item{note}{Note text.}
}
\description{
Append a non-blocking GCAM-China runtime note for end-of-run reporting.
}
\keyword{internal}
