% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_gcam_china_zero_report_vars}
\alias{add_gcam_china_zero_report_vars}
\title{add_gcam_china_zero_report_vars}
\usage{
add_gcam_china_zero_report_vars(report)
}
\description{
Add selected official GCAM-China zero-valued placeholders to the final report
when they are absent.
}
\keyword{internal}
