% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_missing_parent_aggregates}
\alias{add_missing_parent_aggregates}
\title{add_missing_parent_aggregates}
\usage{
add_missing_parent_aggregates(data, template_vars)
}
\description{
Add missing aggregate parent variables for additive IAMC hierarchies.
}
\keyword{internal}
