% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_missing_parent_aggregates_report}
\alias{add_missing_parent_aggregates_report}
\title{add_missing_parent_aggregates_report}
\usage{
add_missing_parent_aggregates_report(report, template_data)
}
\description{
Add missing aggregate parent variables directly to the final wide report.
}
\keyword{internal}
