% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_demand_map}
\alias{ag_demand_map}
\title{ag_demand_map}
\format{
.csv
}
\source{
github
}
\usage{
ag_demand_map
}
\description{
ag maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::ag_demand_map
}
}
\keyword{datasets}
