% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_demand_price_map_v9.1}
\alias{ag_demand_price_map_v9.1}
\title{ag_demand_price_map_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
ag_demand_price_map_v9.1
}
\description{
ag maps compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::ag_demand_price_map_v9.1
}
}
\keyword{datasets}
