% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_prices_map}
\alias{ag_prices_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
ag_prices_map
}
\description{
ag maps
\dontrun{
 library(gcamreport);
 gcamreport::ag_prices_map
}
}
\keyword{datasets}
