% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_crop_production.R
\name{allocate_crop_production}
\alias{allocate_crop_production}
\title{Allocate National Crop Production to Provinces}
\usage{
allocate_crop_production(prj, scenario)
}
\arguments{
\item{prj}{GCAM project object}

\item{scenario}{Scenario name}
}
\value{
Data frame with provincial crop production
}
\description{
Allocates national-level crop production to provinces based on provincial
crop demand as a proxy.
}
