% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_livestock_production.R
\name{allocate_livestock_production}
\alias{allocate_livestock_production}
\title{Allocate National Livestock Production to Provinces}
\usage{
allocate_livestock_production(prj, scenario)
}
\arguments{
\item{prj}{GCAM project object}

\item{scenario}{Scenario name}
}
\value{
Data frame with provincial livestock production
}
\description{
Allocates national-level livestock production to provinces based on provincial
feed crop demand as a proxy.
}
