% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{approx_fun}
\alias{approx_fun}
\title{approx_fun}
\usage{
approx_fun(year, value, rule = 1)

approx_fun(year, value, rule = 1)

approx_fun(year, value, rule = 1)
}
\arguments{
\item{year}{year to consider}

\item{value}{values to extrapolate from}

\item{rule}{number of points to extrapolate}
}
\description{
Interpolation function

Interpolation function

Interpolation function
}
\keyword{internal}
\keyword{utils}
