% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_continents}
\alias{available_continents}
\title{available_continents}
\usage{
available_continents(print = TRUE, GCAM_version = "v7.0")
}
\arguments{
\item{print}{Logical. If TRUE (default), prints the list of available regions' groups. to the console. If FALSE, suppresses the printing and only returns the list.}

\item{GCAM_version}{GCAM version: 'v7.0' (default) or 'v6.0'.}
}
\value{
A vector of character strings representing the names of all available regions' groups. If `print` is TRUE, the function also prints this list to the console.
}
\description{
Retrieves and optionally prints a list of all available continents/regions' groups for the IAMC reporting dataset.
}
\details{
This function provides a list of regions' groups that are available for use in IAMC reporting. By default, it prints this list, but it can also be used to obtain the list programmatically.
}
