% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_final_year}
\alias{available_final_year}
\title{available_final_year}
\format{
vector
}
\source{
local
}
\usage{
available_final_year
}
\description{
gcamreport available final years
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::available_final_year
}
}
\keyword{datasets}
