% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_reporting_years}
\alias{available_reporting_years}
\title{available_reporting_years}
\format{
vector
}
\source{
local
}
\usage{
available_reporting_years
}
\description{
gcamreport reporting available years
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::available_reporting_years
}
}
\keyword{datasets}
