% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildings_en_service}
\alias{buildings_en_service}
\title{buildings_en_service}
\format{
.csv
}
\source{
github
}
\usage{
buildings_en_service
}
\description{
energy services
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::buildings_en_service
}
}
\keyword{datasets}
