% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacity_map}
\alias{capacity_map}
\title{capacity_map}
\format{
.csv
}
\source{
github
}
\usage{
capacity_map
}
\description{
primary, secondary, final energy maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::capacity_map
}
}
\keyword{datasets}
