% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capital_gcam}
\alias{capital_gcam}
\title{capital_gcam}
\format{
.csv
}
\source{
github
}
\usage{
capital_gcam
}
\description{
capital update
\dontrun{
 library(gcamreport);
 gcamreport::capital_gcam
}
}
\keyword{datasets}
