% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capital_gcam_v6.0}
\alias{capital_gcam_v6.0}
\title{capital_gcam_v6.0}
\format{
.csv
}
\source{
github
}
\usage{
capital_gcam_v6.0
}
\description{
capital update compatible with GCAM6.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::capital_gcam_v6.0
}
}
\keyword{datasets}
