% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon_content_v9.1}
\alias{carbon_content_v9.1}
\title{carbon_content_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
carbon_content_v9.1
}
\description{
carbon content compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::carbon_content_v9.1
}
}
\keyword{datasets}
