% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon_densities_vegsoil_v7.0}
\alias{carbon_densities_vegsoil_v7.0}
\title{carbon_densities_vegsoil_v7.0}
\format{
.csv
}
\source{
github
}
\usage{
carbon_densities_vegsoil_v7.0
}
\description{
vegetation and soil carbon densities per basin compatible with GCAM7.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::carbon_densities_vegsoil_v7.0
}
}
\keyword{datasets}
