% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon_densities_vegsoil_vScenarioMIPCMIP7}
\alias{carbon_densities_vegsoil_vScenarioMIPCMIP7}
\title{carbon_densities_vegsoil_vScenarioMIPCMIP7}
\format{
.csv
}
\source{
github
}
\usage{
carbon_densities_vegsoil_vScenarioMIPCMIP7
}
\description{
vegetation and soil carbon densities per basin compatible with GCAMScenarioMIPCMIP7.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::carbon_densities_vegsoil_vScenarioMIPCMIP7
}
}
\keyword{datasets}
