% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cereal_scaler_v9.1}
\alias{cereal_scaler_v9.1}
\title{cereal_scaler_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
cereal_scaler_v9.1
}
\description{
cereal yield and physical land scaler; compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::cereal_scaler_v9.1
}
}
\keyword{datasets}
