% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_gcam}
\alias{cf_gcam}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
cf_gcam
}
\description{
primary, secondary, final energy maps
\dontrun{
 library(gcamreport);
 gcamreport::cf_gcam
}
}
\keyword{datasets}
