% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_gcam_v7.1}
\alias{cf_gcam_v7.1}
\title{cf_gcam_v7.1}
\format{
.csv
}
\source{
github
}
\usage{
cf_gcam_v7.1
}
\description{
primary, secondary, final energy maps compatible with GCAM7.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::cf_gcam_v7.1
}
}
\keyword{datasets}
