% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{check_user_choices_plot}
\alias{check_user_choices_plot}
\title{check_user_choices_plot}
\usage{
check_user_choices_plot(vars, scen, years, reg, grouped)

check_user_choices_plot(vars, scen, years, reg, grouped)
}
\arguments{
\item{scen:}{user's selected scenarios}

\item{years:}{user's selected years}

\item{reg:}{user's selected regions}

\item{vars:}{user's selected variables}

\item{grouped:}{if TRUE, aim to display grouped plot; ungrouped plot otherwise}
}
\description{
Check user's choices to do the plot: at least one scenario, variable, year, and region
must be chosen. In case of 'grouped' plot, all variables must be from the same category.

Check user's choices to do the plot: at least one scenario, variable, year, and region
must be chosen. In case of 'grouped' plot, all variables must be from the same category.
}
\keyword{internal}
