% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2_equivalent}
\alias{CO2_equivalent}
\title{CO2_equivalent}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
CO2_equivalent
}
\description{
CO2 equivalency conversion
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::CO2_equivalent
}
}
\keyword{datasets}
