% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_ets_sector_map}
\alias{co2_ets_sector_map}
\title{co2_ets_sector_map}
\format{
.csv
}
\source{
github
}
\usage{
co2_ets_sector_map
}
\description{
emissions maps
\dontrun{
 library(gcamreport);
 gcamreport::co2_ets_sector_map
}
}
\keyword{datasets}
