% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2_market}
\alias{CO2_market}
\title{CO2_market}
\format{
.csv
}
\source{
github
}
\usage{
CO2_market
}
\description{
new CO2 market
\dontrun{
 library(gcamreport);
 gcamreport::CO2_market
}
}
\keyword{datasets}
