% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2_market_v7.0}
\alias{CO2_market_v7.0}
\title{CO2_market_v7.0}
\format{
.csv
}
\source{
github
}
\usage{
CO2_market_v7.0
}
\description{
new CO2 market compatible with GCAM7.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::CO2_market_v7.0
}
}
\keyword{datasets}
