% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_resource_map_vGCAMChina8.0}
\alias{co2_resource_map_vGCAMChina8.0}
\title{co2_resource_map_vGCAMChina8.0}
\format{
data frame
}
\usage{
co2_resource_map_vGCAMChina8.0
}
\description{
CO2 resource mapping for GCAM-China 8.0
}
\keyword{datasets}
