% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colScaleFuel}
\alias{colScaleFuel}
\title{description}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
colScaleFuel
}
\description{
fuels color to paint
\dontrun{
 library(gcamreport);
 gcamreport::colScaleFuel
}
}
\keyword{datasets}
