% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colScaleTech}
\alias{colScaleTech}
\title{description}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
colScaleTech
}
\description{
techology painting color
\dontrun{
 library(gcamreport);
 gcamreport::colScaleTech
}
}
\keyword{datasets}
