% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_sec_prevsec_weight}
\alias{compute_sec_prevsec_weight}
\title{compute_sec_prevsec_weight}
\usage{
compute_sec_prevsec_weight(dt)
}
\arguments{
\item{dt}{dataset with the following columns: scenario, region, var (reporting variable), year, value}
}
\value{
dataset with `sec_prevsec_weight` column.
}
\description{
An internal function designed to compute the regional weights of a set of variables. The World region
is considered as the annual unit.
}
