% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_19USD_75USD}
\alias{conv_19USD_75USD}
\title{conv_19USD_75USD}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
conv_19USD_75USD
}
\description{
units conversion
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::conv_19USD_75USD
}
}
\keyword{datasets}
