% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_90USD_10USD}
\alias{conv_90USD_10USD}
\title{conv_90USD_10USD}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
conv_90USD_10USD
}
\description{
units conversion
\dontrun{
 library(gcamreport);
 gcamreport::conv_90USD_10USD
}
}
\keyword{datasets}
