% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{conv_EJ_GW}
\alias{conv_EJ_GW}
\title{conv_EJ_GW}
\usage{
conv_EJ_GW(data, cf, EJ, GCAM_version = "v7.0")
}
\arguments{
\item{data}{dataset}

\item{cf}{conversion factor}

\item{EJ}{EJ amount}

\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\description{
Covert EJ to GW
}
\keyword{conversion}
\keyword{internal}
