% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_thousand_million}
\alias{conv_thousand_million}
\title{description}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
conv_thousand_million
}
\description{
units conversion
\dontrun{
 library(gcamreport);
 gcamreport::conv_thousand_million
}
}
\keyword{datasets}
