% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(type)
}
\arguments{
\item{type:}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'}
}
\value{
dataframe with the data from the query
}
\description{
Add nonCO2 large queries
}
