% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{do_collapse_df}
\alias{do_collapse_df}
\title{do_collapse_df}
\usage{
do_collapse_df(basic_data)
}
\arguments{
\item{basic_data:}{dataframe/datatable to be collapsed}
}
\value{
Vector with the collapsed data
}
\description{
Collapse all columns by row of the dataframe and remove NA
}
