% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{do_data_sample}
\alias{do_data_sample}
\title{do_data_sample}
\usage{
do_data_sample(
  sdata,
  sel_scen,
  sel_years,
  sel_cols,
  sel_vars,
  sel_reg,
  basic_reg,
  basic_vars
)

do_data_sample(
  sdata,
  sel_scen,
  sel_years,
  sel_cols,
  sel_vars,
  sel_reg,
  basic_reg,
  basic_vars
)
}
\arguments{
\item{sdata:}{dataset}

\item{sel_scen:}{selected scenario/s}

\item{sel_years:}{selected year/s}

\item{sel_cols:}{selected column/s}

\item{sel_vars:}{selected variables in tree}

\item{sel_reg:}{selected regions in tree}
}
\value{
subseted dataset

subseted dataset
}
\description{
Function to subset a dataset with the user's choices

Function to subset a dataset with the user's choices
}
\keyword{internal}
