% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{do_list}
\alias{do_list}
\title{do_list}
\usage{
do_list(db)
}
\arguments{
\item{db:}{dataframe}
}
\value{
itemized list
}
\description{
Create a list of items with separator to match the style of the 'Variables' column
}
