% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{do_mount_db}
\alias{do_mount_db}
\title{do_mount_db}
\usage{
do_mount_db(data, current_row = 1, current_col = 1, db = NULL, ref)
}
\arguments{
\item{data:}{list of items}

\item{current_row:}{number of the current row}

\item{current_column:}{number of the current column}

\item{db:}{final dataframe}

\item{ref:}{reference dataframe}
}
\value{
itemized dataframe
}
\description{
Build a database by iterating through a list of items using the values of
previously-recorded items to determine which column to place it in.
}
