% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{do_unmount_tree}
\alias{do_unmount_tree}
\title{do_unmount_tree}
\usage{
do_unmount_tree(base_tree, type)
}
\arguments{
\item{base_tree:}{nested list}

\item{type:}{'variables' if nested list is refereed to variables, 'regions'
if it refers to regional  aggregation}
}
\value{
list of "nested" items
}
\description{
Create a list of "nested" items witha separator from a nested list.
}
