% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_gen_map}
\alias{elec_gen_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
elec_gen_map
}
\description{
primary, secondary, final energy maps
\dontrun{
 library(report);
 report::elec_gen_map
}
}
\keyword{datasets}
