% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emissions_list}
\alias{emissions_list}
\title{emissions_list}
\format{
vector
}
\source{
local
}
\usage{
emissions_list
}
\description{
nonCO2 pollutants list
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::emissions_list
}
}
\keyword{datasets}
