% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en_blocks}
\alias{en_blocks}
\title{en_blocks}
\format{
An object of class \code{data.frame} with 31 rows and 8 columns.
}
\usage{
en_blocks
}
\description{
energy variables hierarchy
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::en_blocks
}
}
\keyword{datasets}
