% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en_multiplier_v9.1}
\alias{en_multiplier_v9.1}
\title{en_multiplier_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
en_multiplier_v9.1
}
\description{
energy services compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::en_multiplier_v9.1
}
}
\keyword{datasets}
