% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_prices_map}
\alias{energy_prices_map}
\title{energy_prices_map}
\format{
.csv
}
\source{
github
}
\usage{
energy_prices_map
}
\description{
primary, secondary, final energy maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::energy_prices_map
}
}
\keyword{datasets}
