% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F_GASES}
\alias{F_GASES}
\title{description}
\format{
An object of class \code{character} of length 13.
}
\usage{
F_GASES
}
\description{
ghg emission conversion
\dontrun{
 library(gcamreport);
 gcamreport::F_GASES
}
}
\keyword{datasets}
