% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fig_dir}
\alias{fig_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
fig_dir
}
\description{
figures directory
\dontrun{
 library(gcamreport);
 gcamreport::fig_dir
}
}
\keyword{datasets}
