% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fillScaleFuel}
\alias{fillScaleFuel}
\title{description}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
fillScaleFuel
}
\description{
fuels color to fill
\dontrun{
 library(report);
 report::fillScaleFuel
}
}
\keyword{datasets}
