% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fillScaleTech}
\alias{fillScaleTech}
\title{description}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
fillScaleTech
}
\description{
techology filling color
\dontrun{
 library(gcamreport);
 gcamreport::fillScaleTech
}
}
\keyword{datasets}
