% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_desired_regions}
\alias{filter_desired_regions}
\title{filter_desired_regions}
\usage{
filter_desired_regions(des_reg)
}
\arguments{
\item{des_reg}{vector of the user desired regions}
}
\value{
vector of the desired regions available in the loaded project
}
\description{
Return vector of the desired regions available in the loaded project
}
