% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_loading_regions}
\alias{filter_loading_regions}
\title{filter_loading_regions}
\usage{
filter_loading_regions(data, desired_regions = "All", variable)
}
\arguments{
\item{data}{Dataframe to filter.}

\item{desired_regions}{Vector of regions to include. Defaults to 'All'. To view available regions, run `available_regions()`. The dataset will only include the specified regions.}

\item{variable}{Variable information for the dataset.}
}
\value{
Filtered dataframe.
}
\description{
Filters a GCAM project dataframe by the desired regions.
}
\keyword{internal}
