% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_variables}
\alias{filter_variables}
\title{filter_variables}
\usage{
filter_variables(data)
}
\arguments{
\item{data:}{dataframe to be filtered}
}
\value{
filtered dataframe
}
\description{
Filter the desired regions of a GCAM project
}
