% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_energy_map}
\alias{final_energy_map}
\title{final_energy_map}
\format{
.csv
}
\source{
github
}
\usage{
final_energy_map
}
\description{
primary, secondary, final energy maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::final_energy_map
}
}
\keyword{datasets}
