% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_energy_map_v6.0}
\alias{final_energy_map_v6.0}
\title{final_energy_map_v6.0}
\format{
.csv
}
\source{
github
}
\usage{
final_energy_map_v6.0
}
\description{
primary, secondary, final energy maps compatible with GCAM6.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::final_energy_map_v6.0
}
}
\keyword{datasets}
