% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel.color}
\alias{fuel.color}
\title{description}
\format{
An object of class \code{character} of length 9.
}
\usage{
fuel.color
}
\description{
fuels color
\dontrun{
 library(gcamreport);
 gcamreport::fuel.color
}
}
\keyword{datasets}
