% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{gather_map}
\alias{gather_map}
\title{gather_map}
\usage{
gather_map(df)

gather_map(df)

gather_map(df)
}
\description{
Formats all maps into a long table

Formats all maps into a long table

Formats all maps into a long table
}
\keyword{internal}
