% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_years}
\alias{gather_years}
\title{gather_years}
\usage{
gather_years(
  d,
  value_col = "value",
  year_pattern = "^(1|2)[0-9]{3}$",
  na.rm = FALSE
)
}
\description{
Formats multiple years into a long-format table.
Source: gcamdata
}
\keyword{internal}
