% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcam_years}
\alias{gcam_years}
\title{gcam_years}
\format{
vector
}
\source{
local
}
\usage{
gcam_years
}
\description{
GCAM model years
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::gcam_years
}
}
\keyword{datasets}
