% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_ppp_2021_region}
\alias{gdp_ppp_2021_region}
\title{gdp_ppp_2021_region}
\format{
.csv
}
\source{
World Bank WDI NY.GDP.MKTP.PP.KD; SSP database 2025 where the World Bank has no value
}
\usage{
gdp_ppp_2021_region
}
\description{
GDP at PPP in 2021 by GCAM region (billion USD 2010), the target GDP|PPP is calibrated to.
Built from inst/extdata/mappings/common/GDP_PPP_2021_ctry.csv.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::gdp_ppp_2021_region
}
}
\keyword{datasets}
