% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_ag_demand}
\alias{get_ag_demand}
\title{get_ag_demand}
\usage{
get_ag_demand()

get_ag_demand()

get_ag_demand()
}
\value{
ag_demand_clean global variable

ag_demand_clean global variable

ag_demand_clean global variable
}
\description{
Get agricultural demand.

Get agricultural demand.

Get agricultural demand.
}
\keyword{ag}
\keyword{internal}
