% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_ag_prices_wld_tmp}
\alias{get_ag_prices_wld_tmp}
\title{get_ag_prices_wld_tmp}
\usage{
get_ag_prices_wld_tmp()

get_ag_prices_wld_tmp()

get_ag_prices_wld_tmp()
}
\value{
ag_prices_wld global variable

ag_prices_wld global variable

ag_prices_wld global variable
}
\description{
Get ag price index.

Get ag price index.

Get ag price index.
}
\keyword{ag}
\keyword{internal}
\keyword{tmp}
