% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ag_production}
\alias{get_ag_production}
\title{get_ag_production}
\usage{
get_ag_production()
}
\value{
ag_production_clean global variable
}
\description{
Get agricultural production.
}
\keyword{ag}
\keyword{rgcam::getQuery}
