% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cdr_lulucf}
\alias{get_cdr_lulucf}
\title{get_cdr_lulucf}
\usage{
get_cdr_lulucf(GCAM_version = "v7.1")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.1' (default), 'v7.2', 'v7.0'.}
}
\value{
`cdr_lulucf` global variable.
}
\description{
Get carbon dioxide removal from LULUCF.
}
\keyword{co2}
\keyword{internal}
