% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_co2_concentration}
\alias{get_co2_concentration}
\title{get_co2_concentration}
\usage{
get_co2_concentration()
}
\value{
co2_concentration_clean global variable
}
\description{
Get World's CO2 concentration query.
}
\keyword{co2}
\keyword{getQuery}
