% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2_ets}
\alias{get_co2_ets}
\title{get_co2_ets}
\usage{
get_co2_ets()

get_co2_ets()

get_co2_ets()
}
\value{
co2_ets_by reg and co2_ets_bysec global variables

co2_ets_by reg and co2_ets_bysec global variables

co2_ets_by reg and co2_ets_bysec global variables
}
\description{
Get World's CO2 ETS emissions query.

Get World's CO2 ETS emissions query.

Get World's CO2 ETS emissions query.
}
\keyword{co2}
\keyword{internal}
