% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_iron_steel}
\alias{get_co2_iron_steel}
\title{get_co2_iron_steel}
\usage{
get_co2_iron_steel(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
co2_tech_ironsteel global variable
}
\description{
Get iron and steel CO2 emissions.
}
\keyword{co2}
\keyword{internal}
\keyword{iron}
\keyword{steel}
