% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2_price}
\alias{get_co2_price}
\title{get_co2_price}
\usage{
get_co2_price()

get_co2_price()

get_co2_price()
}
\value{
co2_price_clean global variable

co2_price_clean global variable

co2_price_clean global variable
}
\description{
Get co2 price.

Get co2 price.

Get co2 price.
}
\keyword{co2}
\keyword{internal}
