% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price_share_byreg}
\alias{get_co2_price_share_byreg}
\title{Get CO2 Price Share}
\usage{
get_co2_price_share_byreg(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
Global variable `co2_price_share_byreg` containing CO2 price shares by region.
}
\description{
Retrieves the CO2 price share of each region or sector compared to the total CO2 price.
}
\keyword{co2}
\keyword{internal}
\keyword{tmp}
