% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_co2_tech_nobio_tmp}
\alias{get_co2_tech_nobio_tmp}
\title{get_co2_tech_nobio_tmp}
\usage{
get_co2_tech_nobio_tmp()
}
\value{
co2_tech_nobio global variable
}
\description{
Get no bio CO2 emissions query by sector and techonolgy.
}
\keyword{co2}
\keyword{getQuery}
\keyword{tmp}
