% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_capacity_add}
\alias{get_elec_capacity_add}
\title{get_elec_capacity_add}
\usage{
get_elec_capacity_add()
}
\value{
elec_capacity_add_clean global variable
}
\description{
Calculate added total capacity
}
\keyword{capacity}
\keyword{process}
