% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_capacity_add_tmp}
\alias{get_elec_capacity_add_tmp}
\title{get_elec_capacity_add_tmp}
\usage{
get_elec_capacity_add_tmp()
}
\value{
`elec_capacity_add``` global variable
}
\description{
Calculate added total capacity.
}
\keyword{capacity}
\keyword{internal}
\keyword{process}
\keyword{tmp}
