% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_cf_tmp}
\alias{get_elec_cf_tmp}
\title{get_elec_cf_tmp}
\usage{
get_elec_cf_tmp()
}
\value{
elec_cf global variable
}
\description{
Calculate future capacity using GCAM
}
\keyword{capacity}
\keyword{process}
\keyword{tmp}
