% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_investment}
\alias{get_elec_investment}
\title{get_elec_investment}
\usage{
get_elec_investment(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
elec_investment_clean global variable
}
\description{
Calculate electricity investment = annual capacity additions * capital costs
}
\keyword{capital}
\keyword{internal}
\keyword{process}
