% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_energy_price_fragmented}
\alias{get_energy_price_fragmented}
\title{get_energy_price_fragmented}
\usage{
get_energy_price_fragmented()

get_energy_price_fragmented()

get_energy_price_fragmented()
}
\value{
energy_price_fragmented global variable

energy_price_fragmented global variable

energy_price_fragmented global variable
}
\description{
Get energy prices fragmented, join by region since price is different.

Get energy prices fragmented, join by region since price is different.

Get energy prices fragmented, join by region since price is different.
}
\keyword{internal}
\keyword{prices}
\keyword{process}
