% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_energy_price_tmp}
\alias{get_energy_price_tmp}
\title{get_energy_price_tmp}
\usage{
get_energy_price_tmp()
}
\value{
energy_price global variable
}
\description{
Bind regional oil, gas, coal prices with other energy prices
}
\keyword{getQuery}
\keyword{price}
\keyword{tmp}
