% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_energy_service_buildings}
\alias{get_energy_service_buildings}
\title{get_energy_service_buildings}
\usage{
get_energy_service_buildings()

get_energy_service_buildings()

get_energy_service_buildings()
}
\value{
energy_service_buildings_clean global variable

energy_service_buildings_clean global variable

energy_service_buildings_clean global variable
}
\description{
Get ES buildings.

Get ES buildings.

Get ES buildings.
}
\keyword{energy}
\keyword{internal}
