% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_energy_trade}
\alias{get_energy_trade}
\title{get_energy_trade}
\usage{
get_energy_trade()
}
\value{
energy_trade_clean global variable
}
\description{
Get energy trade supply.
}
\keyword{energy}
\keyword{getQuery}
\keyword{tmp}
