% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_energy_trade_prod}
\alias{get_energy_trade_prod}
\title{get_energy_trade_prod}
\usage{
get_energy_trade_prod()
}
\value{
energy_trade_prod global variable
}
\description{
Get energy trade.
}
\keyword{energy}
\keyword{getQuery}
