% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_trade_tmp}
\alias{get_energy_trade_tmp}
\title{get_energy_trade_tmp}
\usage{
get_energy_trade_tmp()
}
\value{
energy_trade_supply global variable
}
\description{
Get energy trade supply. Query used to compute other variables.
}
\keyword{energy}
\keyword{tmp}
