% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fe_sector_tmp}
\alias{get_fe_sector_tmp}
\title{get_fe_sector_tmp}
\usage{
get_fe_sector_tmp()
}
\value{
fe_sector global variable
}
\description{
Get final energy demand by sector
}
\keyword{energy}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
\keyword{tmp}
