% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_food_weights}
\alias{get_food_weights}
\title{get_food_weights}
\usage{
get_food_weights(GCAM_version = "v8.2")
}
\arguments{
\item{GCAM_version}{Name of the GCAM compatible version. Run `available_GCAM_versions()` to see the list of supported options.}
}
\value{
`food_weights` and `food_wld_weights` global variables.
}
\description{
Get food items weighted by demand. By region and global.
}
\keyword{ag}
\keyword{internal}
