% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_forcing}
\alias{get_forcing}
\title{get_forcing}
\usage{
get_forcing()

get_forcing()

get_forcing()
}
\value{
forcing_clean global variable

forcing_clean global variable

forcing_clean global variable
}
\description{
Get World's forcing query.

Get World's forcing query.

Get World's forcing query.
}
\keyword{forcing}
\keyword{internal}
