% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_capital_gcam}
\alias{get_gcam_china_capital_gcam}
\title{get_gcam_china_capital_gcam}
\usage{
get_gcam_china_capital_gcam()
}
\description{
Load the official GCAM-China electricity capital cost tables.
}
\keyword{internal}
