% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_cf_province}
\alias{get_gcam_china_cf_province}
\title{get_gcam_china_cf_province}
\usage{
get_gcam_china_cf_province()
}
\description{
Load the official GCAM-China province-specific renewable capacity factors.
}
\keyword{internal}
