% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_cf_tech_mapping}
\alias{get_gcam_china_cf_tech_mapping}
\title{get_gcam_china_cf_tech_mapping}
\usage{
get_gcam_china_cf_tech_mapping()
}
\description{
Load the official GCAM-China mapping between GCAM electricity technologies
and province-specific capacity factor adjustment categories.
}
\keyword{internal}
