% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_co2_resource_map}
\alias{get_gcam_china_co2_resource_map}
\title{get_gcam_china_co2_resource_map}
\usage{
get_gcam_china_co2_resource_map()
}
\description{
Load the GCAM-China CO2 resource production mapping.
}
\keyword{internal}
