% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_co2_sector_map}
\alias{get_gcam_china_co2_sector_map}
\title{get_gcam_china_co2_sector_map}
\usage{
get_gcam_china_co2_sector_map()
}
\description{
Load the official GCAM-China CO2 sector map.
}
\keyword{internal}
