% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_elec_gen_map}
\alias{get_gcam_china_elec_gen_map}
\title{get_gcam_china_elec_gen_map}
\usage{
get_gcam_china_elec_gen_map()
}
\description{
Load the official GCAM-China electricity generation map used for
generation, capacity and investment reporting.
}
\keyword{internal}
