% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_final_energy_map}
\alias{get_gcam_china_final_energy_map}
\title{get_gcam_china_final_energy_map}
\usage{
get_gcam_china_final_energy_map()
}
\description{
Load the official GCAM-China final energy mapping.
}
\keyword{internal}
