% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_iea_capacity_2023}
\alias{get_gcam_china_iea_capacity_2023}
\title{get_gcam_china_iea_capacity_2023}
\usage{
get_gcam_china_iea_capacity_2023()
}
\description{
Load the official GCAM-China reference IEA WEO 2023 capacity file.
}
\keyword{internal}
