% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_provinces}
\alias{get_gcam_china_provinces}
\title{get_gcam_china_provinces}
\usage{
get_gcam_china_provinces()
}
\description{
Return the province regions used by GCAM-China reporting.
}
\keyword{internal}
