% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_transport_final_en_map}
\alias{get_gcam_china_transport_final_en_map}
\title{get_gcam_china_transport_final_en_map}
\usage{
get_gcam_china_transport_final_en_map()
}
\description{
Load the official GCAM-China transport final energy mapping.
}
\keyword{internal}
