% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gcam_china_weo_mapping}
\alias{get_gcam_china_weo_mapping}
\title{get_gcam_china_weo_mapping}
\usage{
get_gcam_china_weo_mapping()
}
\description{
Load the official GCAM-China mapping between IAMC electricity variables and
IEA WEO capacity categories.
}
\keyword{internal}
