% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_ppp}
\alias{get_gdp_ppp}
\title{get_gdp_ppp}
\usage{
get_gdp_ppp(GCAM_version = "v8.2")
}
\arguments{
\item{GCAM_version}{Name of the GCAM compatible version. Run `available_GCAM_versions()` to see the list of supported options.}
}
\value{
`GDP_PPP_clean`, `GDP_PPP_pc_growth_clean`, and `GDP_PPP_pc_oecd_share_clean` global variables.
}
\description{
Retrieves GDP (PPP) data, computes regional GDP and annual GDPpc growth rate, and converts units to 10 USD.
Each region is scaled so that 2021 reproduces GDP at PPP from the World Bank (`gdp_ppp_2021_region`).
}
\keyword{econ}
\keyword{internal}
