% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_ghg}
\alias{get_ghg}
\title{get_ghg}
\usage{
get_ghg()

get_ghg()

get_ghg()
}
\value{
ghg_clean global variable

ghg_clean global variable

ghg_clean global variable
}
\description{
Get total GHG emissions.

Get total GHG emissions.

Get total GHG emissions.
}
\keyword{ghg}
\keyword{internal}
