% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ghg_sector}
\alias{get_ghg_sector}
\title{get_ghg_sector}
\usage{
get_ghg_sector(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
ghg_sector_clean global variable
}
\description{
Get sectorial GHG emissions.
}
\keyword{ghg}
\keyword{internal}
