% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_gov_revenue}
\alias{get_gov_revenue}
\title{get_gov_revenue}
\usage{
get_gov_revenue()

get_gov_revenue()

get_gov_revenue()
}
\value{
gov_revenue_all global variable

gov_revenue_all global variable

gov_revenue_all global variable
}
\description{
Get overall carbon revenue.

Get overall carbon revenue.

Get overall carbon revenue.
}
\keyword{internal}
\keyword{revenue}
