% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gov_revenue_all}
\alias{get_gov_revenue_all}
\title{get_gov_revenue_all}
\usage{
get_gov_revenue_all()
}
\value{
gov_revenue_all global variable
}
\description{
Get overall carbon revenue.
}
\keyword{getQuery}
\keyword{revenue}
