% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gov_revenue_sector}
\alias{get_gov_revenue_sector}
\title{get_gov_revenue_sector}
\usage{
get_gov_revenue_sector()
}
\value{
gov_revenue_sector global variable
}
\description{
Get overall carbon revenue.
}
\keyword{internal}
\keyword{revenue}
\keyword{rgcam::getQuery}
