% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam2gains_industry.R
\name{get_industry_output}
\alias{get_industry_output}
\title{Get Industry Primary Output (for validation)}
\usage{
get_industry_output(prj, scenario)
}
\arguments{
\item{prj}{GCAM project object}

\item{scenario}{Scenario name}
}
\value{
Data frame with industry primary output
}
\description{
Get Industry Primary Output (for validation)
}
