% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_clean}
\alias{get_iron_steel_clean}
\title{get_iron_steel_clean}
\usage{
get_iron_steel_clean()
}
\value{
`iron_steel_clean` global variable
}
\description{
Retrieve iron steel imports & exports
}
\keyword{industry}
\keyword{internal}
