% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_imports}
\alias{get_iron_steel_imports}
\title{get_iron_steel_imports}
\usage{
get_iron_steel_imports()
}
\value{
iron_steel_imports global variable
}
\description{
Get iron steel imports
}
\keyword{industry}
\keyword{internal}
\keyword{tmp}
