% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_prod_tech}
\alias{get_iron_steel_prod_tech}
\title{get_iron_steel_prod_tech}
\usage{
get_iron_steel_prod_tech(GCAM_version = "v7.1")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.1' (default), 'v7.2', 'v7.0'.}
}
\value{
`iron_steel_prod_tech_clean` global variable
}
\description{
Retrieve iron and steel production by technology.
}
\keyword{industry}
\keyword{internal}
