% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_land}
\alias{get_land}
\title{get_land}
\usage{
get_land()

get_land()

get_land()
}
\value{
land_clean global variable

land_clean global variable

land_clean global variable
}
\description{
Get land use area.

Get land use area.

Get land use area.
}
\keyword{ag}
\keyword{internal}
