% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_population}
\alias{get_population}
\title{get_population}
\usage{
get_population()

get_population()

get_population()
}
\value{
population_clean global variable

population_clean global variable

population_clean global variable
}
\description{
Get population query and change units to [million].

Get population query and change units to [million].

Get population query and change units to [million].
}
\keyword{internal}
\keyword{population}
