% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_prices_subsector}
\alias{get_prices_subsector}
\title{get_prices_subsector}
\usage{
get_prices_subsector()

get_prices_subsector()

get_prices_subsector()
}
\value{
prices_subsector global variable

prices_subsector global variable

prices_subsector global variable
}
\description{
Get energy prices - primary, secondary, and final

Get energy prices - primary, secondary, and final

Get energy prices - primary, secondary, and final
}
\keyword{internal}
\keyword{prices}
