% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_primary_energy_electricity}
\alias{get_primary_energy_electricity}
\title{get_primary_energy_electricity}
\usage{
get_primary_energy_electricity(GCAM_version = "v7.1")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.1' (default), 'v7.2', 'v7.0'.}
}
\value{
`primary_energy_electricity_clean` global variable.
}
\description{
Retrieve primary energy from electricity generation by fuel and CCS.
}
\keyword{energy}
\keyword{internal}
