% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_regional_emission}
\alias{get_regional_emission}
\title{get_regional_emission}
\usage{
get_regional_emission()

get_regional_emission()

get_regional_emission()
}
\value{
regional_emission global variable

regional_emission global variable

regional_emission global variable
}
\description{
Compute regional nonCO2 emission: regional production * nonCO2 coef.

Compute regional nonCO2 emission: regional production * nonCO2 coef.

Compute regional nonCO2 emission: regional production * nonCO2 coef.
}
\keyword{internal}
\keyword{nonco2}
