% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regions_en_weight_tmp}
\alias{get_regions_en_weight_tmp}
\title{get_regions_en_weight_tmp}
\usage{
get_regions_en_weight_tmp()
}
\value{
`region_en_weight` global variable
}
\description{
Retrieve regions final energy weights to compute carbon price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
