% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regions_tmp}
\alias{get_regions_tmp}
\title{get_regions_tmp}
\usage{
get_regions_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
regions global variable
}
\description{
Get regions to compute carbon price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
