% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_resource_investment}
\alias{get_resource_investment}
\title{get_resource_investment}
\usage{
get_resource_investment()

get_resource_investment()

get_resource_investment()
}
\value{
resource_investment_clean global variable

resource_investment_clean global variable

resource_investment_clean global variable
}
\description{
Calculate investment of resource production

Calculate investment of resource production

Calculate investment of resource production
}
\keyword{internal}
\keyword{investment}
\keyword{process}
