% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_runtime_convert}
\alias{get_runtime_convert}
\title{get_runtime_convert}
\usage{
get_runtime_convert(GCAM_version = "v7.1")
}
\description{
Return conversion constants, correcting known runtime issues for
`vGCAMChina7.1`.
}
\keyword{internal}
