% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_se_gen_tech}
\alias{get_se_gen_tech}
\title{get_se_gen_tech}
\usage{
get_se_gen_tech()
}
\value{
se_gen_tech_clean global variable
}
\description{
Get other secondary energy production
}
\keyword{energy}
\keyword{getQuery}
