% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam2gains_industry.R
\name{get_steel_production}
\alias{get_steel_production}
\title{Get Iron and Steel Production by Technology}
\usage{
get_steel_production(prj, scenario)
}
\arguments{
\item{prj}{GCAM project object}

\item{scenario}{Scenario name}
}
\value{
Data frame with iron and steel production
}
\description{
Get Iron and Steel Production by Technology
}
