% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_temperature}
\alias{get_temperature}
\title{get_temperature}
\usage{
get_temperature()

get_temperature()

get_temperature()
}
\value{
global_temp_clean global variable

global_temp_clean global variable

global_temp_clean global variable
}
\description{
Get World's mean temperature query.

Get World's mean temperature query.

Get World's mean temperature query.
}
\keyword{internal}
\keyword{temperature}
