% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_total_co2_emissions}
\alias{get_total_co2_emissions}
\title{get_total_co2_emissions}
\usage{
get_total_co2_emissions()
}
\value{
tot_co2_clean global variable
}
\description{
Compute total CO2 emission.
}
\keyword{co2}
\keyword{internal}
\keyword{process}
