% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_total_revenue}
\alias{get_total_revenue}
\title{get_total_revenue}
\usage{
get_total_revenue(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
total_revenue global variable
}
\description{
Compute total revenue: total production * global price.
}
\keyword{internal}
\keyword{revenue}
