% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_transmission_invest}
\alias{get_transmission_invest}
\title{get_transmission_invest}
\usage{
get_transmission_invest(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
transmission_invest_clean global variable
}
\description{
Calculate investment electricity transmission and distribution
scale 2020 number - average of other model results from Mcollion et al. 2018
Need to convert 2015 to 2010 $
}
\keyword{internal}
\keyword{investment}
\keyword{process}
