% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam2gains_transport.R
\name{get_transport_service_output}
\alias{get_transport_service_output}
\title{Get Transport Service Output (for validation)}
\usage{
get_transport_service_output(prj, scenario)
}
\arguments{
\item{prj}{GCAM project object}

\item{scenario}{Scenario name}
}
\value{
Data frame with transport service output
}
\description{
Get Transport Service Output (for validation)
}
