% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHG_gases}
\alias{GHG_gases}
\title{GHG_gases}
\format{
An object of class \code{character} of length 20.
}
\usage{
GHG_gases
}
\description{
ghg emission conversion
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::GHG_gases
}
}
\keyword{datasets}
