% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghg_GWP_AR4}
\alias{ghg_GWP_AR4}
\title{ghg_GWP_AR4}
\format{
.csv
}
\source{
local
}
\usage{
ghg_GWP_AR4
}
\description{
ghg adjuster - uses 100-yr GWPs from AR4 in GCAM unit
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::ghg_GWP_AR4
}
}
\keyword{datasets}
