% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_vet_values}
\alias{global_vet_values}
\title{global_vet_values}
\format{
.csv
}
\source{
local
}
\usage{
global_vet_values
}
\description{
vetting values to test
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::global_vet_values
}
}
\keyword{datasets}
