% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWP_adjuster}
\alias{GWP_adjuster}
\title{GWP_adjuster}
\format{
.csv
}
\source{
local
}
\usage{
GWP_adjuster
}
\description{
ghg adjuster
\dontrun{
 library(gcamreport);
 gcamreport::GWP_adjuster
}
}
\keyword{datasets}
