% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
<<<<<<<< HEAD:man/ghg_GWP_AR4.Rd
\name{ghg_GWP_AR4}
\alias{ghg_GWP_AR4}
\title{ghg_GWP_AR4}
========
\name{GWP_adjuster_v7.0}
\alias{GWP_adjuster_v7.0}
\title{GWP_adjuster_v7.0}
>>>>>>>> dev_newstructure:man/GWP_adjuster_v7.0.Rd
\format{
.csv
}
\source{
local
}
\usage{
<<<<<<<< HEAD:man/ghg_GWP_AR4.Rd
ghg_GWP_AR4
}
\description{
ghg adjuster - uses 100-yr GWPs from AR4 in GCAM unit
========
GWP_adjuster_v7.0
}
\description{
ghg adjuster compatible with GCAM7.0.
>>>>>>>> dev_newstructure:man/GWP_adjuster_v7.0.Rd
}
\examples{
\dontrun{
library(gcamreport)
<<<<<<<< HEAD:man/ghg_GWP_AR4.Rd
gcamreport::ghg_GWP_AR4
========
gcamreport::GWP_adjuster_v7.0
>>>>>>>> dev_newstructure:man/GWP_adjuster_v7.0.Rd
}
}
\keyword{datasets}
