% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handle_warning}
\alias{handle_warning}
\title{handle_warning}
\usage{
handle_warning(mapping_name1, mapping_name2 = NULL, query_name = NULL)
}
\arguments{
\item{mapping_name1}{The name of the mapping file 1.}

\item{mapping_name2}{The name of the mapping file 2.}

\item{query_name}{The name of the query.}
}
\value{
A warning is issued, and the user's decision (manual check or continuation) is captured.
}
\description{
An internal function designed to manage mismatches between mapping files and queries.
It prompts the user for input, offering the option to manually check the issue or to continue
with the current process.
}
