% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hr_per_yr}
\alias{hr_per_yr}
\title{description}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
hr_per_yr
}
\description{
units conversion
\dontrun{
 library(report);
 report::hr_per_yr
}
}
\keyword{datasets}
