% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iamc_dir}
\alias{iamc_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
iamc_dir
}
\description{
main directory
\dontrun{
 library(report);
 report::iamc_dir
}
}
\keyword{datasets}
