% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iea_capacity}
\alias{iea_capacity}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
iea_capacity
}
\description{
iea 2019 capacity
\dontrun{
 library(gcamreport);
 gcamreport::iea_capacity
}
}
\keyword{datasets}
