% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_gcam_china_run_notes}
\alias{init_gcam_china_run_notes}
\title{init_gcam_china_run_notes}
\usage{
init_gcam_china_run_notes()
}
\description{
Initialize the GCAM-China run note collector.
}
\keyword{internal}
