% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{investment}
\alias{investment}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
investment
}
\description{
capital investment
\dontrun{
 library(gcamreport);
 gcamreport::investment
}
}
\keyword{datasets}
