% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{investment_v9.1}
\alias{investment_v9.1}
\title{investment_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
investment_v9.1
}
\description{
capital investment compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::investment_v9.1
}
}
\keyword{datasets}
