% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kyoto_sector_map}
\alias{kyoto_sector_map}
\title{kyoto_sector_map}
\format{
.csv
}
\source{
github
}
\usage{
kyoto_sector_map
}
\description{
emissions maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::kyoto_sector_map
}
}
\keyword{datasets}
