% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kyoto_sector_map_vGCAMChina8.0}
\alias{kyoto_sector_map_vGCAMChina8.0}
\title{kyoto_sector_map_vGCAMChina8.0}
\format{
data frame
}
\usage{
kyoto_sector_map_vGCAMChina8.0
}
\description{
Kyoto sector mapping for GCAM-China 8.0
}
\keyword{datasets}
