% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_use_map}
\alias{land_use_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
land_use_map
}
\description{
ag maps
\dontrun{
 library(gcamreport);
 gcamreport::land_use_map
}
}
\keyword{datasets}
