% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_historical_year}
\alias{last_historical_year}
\title{last_historical_year}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
last_historical_year
}
\description{
last historical years
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::last_historical_year
}
}
\keyword{datasets}
