% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{left_join_strict}
\alias{left_join_strict}
\title{left_join_strict}
\usage{
left_join_strict(
  left_df,
  right_df,
  by = NULL,
  by_message = by,
  mapping = "",
  ...
)
}
\arguments{
\item{left_df}{A data frame. The left dataset in the join.}

\item{right_df}{A data frame. The right dataset in the join.}

\item{by}{A character vector of variables to join by. If `NULL`, the function will use all common variables.}

\item{by_message}{A character vector of variables to join by to output the ERROR message, if necessay. If `NULL`, the function will use the variables defined in the `by` parameter.}

\item{mapping}{Optional. Mapping name to be displayed in case of ERROR.}

\item{...}{Additional arguments passed to `dplyr::left_join()`.}
}
\value{
A data frame resulting from the left join. If any rows in `left_df` do not have matching keys in `right_df`, an error is thrown.
}
\description{
A restrictive version of \code{\link{left_join}} that ensures that all keys in the left dataset have corresponding matches in the right dataset.
If any rows in the left dataset do not have matching keys in the right dataset, the function will throw an error.
}
