% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam2gains_buildings.R
\name{load_gcam2gains_mapping}
\alias{load_gcam2gains_mapping}
\title{Load GCAM2GAINS Mapping Files}
\usage{
load_gcam2gains_mapping(mapping_name)
}
\arguments{
\item{mapping_name}{Name of the mapping file (without .csv extension)}
}
\value{
Data frame with mapping information
}
\description{
Load GCAM2GAINS Mapping Files
}
