% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_project}
\alias{load_project}
\title{load_project}
\usage{
load_project(prj_name)
}
\arguments{
\item{prj_name:}{name of the project}
}
\value{
loaded project into global environment
}
\description{
Load specified project into the global environment
}
