% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_dir}
\alias{map_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
map_dir
}
\description{
mapping directory
\dontrun{
 library(report);
 report::map_dir
}
}
\keyword{datasets}
