% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco2_emis_sector_map}
\alias{nonco2_emis_sector_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
nonco2_emis_sector_map
}
\description{
emissions maps
\dontrun{
 library(report);
 report::nonco2_emis_sector_map
}
}
\keyword{datasets}
