% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco2_emis_sector_map_v9.1}
\alias{nonco2_emis_sector_map_v9.1}
\title{nonco2_emis_sector_map_v9.1}
\format{
.csv
}
\source{
github
}
\usage{
nonco2_emis_sector_map_v9.1
}
\description{
emissions maps compatible with GCAM9.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::nonco2_emis_sector_map_v9.1
}
}
\keyword{datasets}
