% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco2_emis_sector_map_vGCAMChina8.0}
\alias{nonco2_emis_sector_map_vGCAMChina8.0}
\title{nonco2_emis_sector_map_vGCAMChina8.0}
\format{
data frame
}
\usage{
nonco2_emis_sector_map_vGCAMChina8.0
}
\description{
Non-CO2 emissions sector mapping for GCAM-China 8.0
}
\keyword{datasets}
