% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco2_emissions_list_v9.1}
\alias{nonco2_emissions_list_v9.1}
\title{nonco2_emissions_list_v9.1}
\format{
vector
}
\source{
local
}
\usage{
nonco2_emissions_list_v9.1
}
\description{
nonCO2 pollutants list
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::nonco2_emissions_list_v9.1
}
}
\keyword{datasets}
