% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_gcam_china_power_technology}
\alias{normalize_gcam_china_power_technology}
\title{normalize_gcam_china_power_technology}
\usage{
normalize_gcam_china_power_technology(technology)
}
\description{
Normalize GCAM-China segmented power technology names to the standard
technology names used in capacity mappings.
}
\keyword{internal}
