% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_dir}
\alias{out_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
out_dir
}
\description{
output directory
\dontrun{
 library(gcamreport);
 gcamreport::out_dir
}
}
\keyword{datasets}
