% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production_map}
\alias{production_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
production_map
}
\description{
primary, secondary, final energy maps
\dontrun{
 library(report);
 report::production_map
}
}
\keyword{datasets}
