% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_general_v7.0}
\alias{queries_general_v7.0}
\title{queries_general_v7.0}
\format{
vector
}
\source{
local
}
\usage{
queries_general_v7.0
}
\description{
gcamreport queries compatible with GCAM 7.0 version compatible with GCAM7.0.
Contain all queries except for nonCO2 queries
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::queries_general_v7.0
}
}
\keyword{datasets}
