% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_general_vGCAMChina8.0}
\alias{queries_general_vGCAMChina8.0}
\title{queries_general_vGCAMChina8.0}
\format{
character vector
}
\usage{
queries_general_vGCAMChina8.0
}
\description{
General queries for GCAM-China 8.0
}
\keyword{datasets}
