% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_nonCO2_v7.1}
\alias{queries_nonCO2_v7.1}
\title{queries_nonCO2_v7.1}
\format{
vector
}
\source{
local
}
\usage{
queries_nonCO2_v7.1
}
\description{
gcamreport nonCO2 query compatible with GCAM 7.1 version compatible with GCAM7.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::queries_nonCO2_v7.1
}
}
\keyword{datasets}
