% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_nonCO2_vEurope8.7}
\alias{queries_nonCO2_vEurope8.7}
\title{queries_nonCO2_vEurope8.7}
\format{
vector
}
\source{
local
}
\usage{
queries_nonCO2_vEurope8.7
}
\description{
gcamreport nonCO2 query compatible with GCAM Europe 8.7 version.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::queries_nonCO2_vEurope8.7
}
}
\keyword{datasets}
