% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_queries}
\alias{read_queries}
\title{read_queries}
\usage{
read_queries()
}
\value{
load variable
}
\description{
Main function. Interacts with the user to select the desired variables for the report,
loads them, saves them in an external output, runs the verifications, and informs the
user about the success of the whole process.
}
\keyword{internal}
