% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_cont}
\alias{reg_cont}
\title{description}
\format{
.csv
}
\source{
local
}
\usage{
reg_cont
}
\description{
mapping between regions and continents
\dontrun{
 library(gcamreport);
 gcamreport::reg_cont
}
}
\keyword{datasets}
