% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporting_columns}
\alias{reporting_columns}
\title{description}
\format{
An object of class \code{character} of length 25.
}
\usage{
reporting_columns
}
\description{
reporting columns
\dontrun{
 library(gcamreport);
 gcamreport::reporting_columns
}
}
\keyword{datasets}
