% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporting_years}
\alias{reporting_years}
\title{reporting_years}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
reporting_years
}
\description{
reporting years
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::reporting_years
}
}
\keyword{datasets}
