% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reporting_years_v6.0}
\alias{reporting_years_v6.0}
\title{reporting_years_v6.0}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
reporting_years_v6.0
}
\description{
reporting years compatible with GCAM6.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::reporting_years_v6.0
}
}
\keyword{datasets}
