% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run}
\alias{run}
\title{run}
\usage{
run(
  project_path = NULL,
  db_path = NULL,
  db_name = NULL,
  prj_name = NULL,
  scenarios = NULL,
  final_year = 2100,
  desired_variables = "All",
  desired_regions = "All",
  desired_continents = "All",
  save_output = TRUE,
  file_name = NULL,
  launch_ui = TRUE
)
}
\arguments{
\item{project_path:}{full path of the project with the project name. Possible extensions: .dat and .proj.}

\item{db_path:}{full path of the database.}

\item{db_name:}{name of the database.}

\item{prj_name:}{name of the project.}

\item{scenarios:}{name of the scenarios to be considered.}

\item{final_year:}{final year of the data. By default = 2100. ATENTION: final_year must be at least 2025.}

\item{desired_variables:}{desired variables to have in the report. Considered 'All' by default.
Otherwise, specify a vector with all the desired options. To know all possible options, run `available_variables()`.
In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3"}

\item{desired_regions:}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
In case the project dataset needs to be created, it will be produced with only the specified regions.}

\item{desired_continents:}{desired continents/regions' groups to consider. By default, 'All'. Otherwise, specify a vector with all
the considered continents/regions' groups. To know all possibilities, run `available_continents()`. ATTENTION: the considered
continents/regions' groups will make up "World". In case the project dataset needs to be created, it will be produced with only the
specified continents/regions' groups.}

\item{save_output:}{if TRUE, save reporting data in CSV and XLSX formats. If FALSE, do not save data. If equals 'CSV' or 'XLSX',
data saved only in the specified format.}

\item{file_name:}{file path and name of the saved data. Not used if data not saved. By default, saved in the same directory and with
the same name than the specified project_path, with 'iamc_report' tag. CSV and XLSX output. In case of specifing the path, do not
introduce the extension, it will be automatically added.}

\item{launch_ui:}{if TRUE, launch UI, Do not launch UI otherwise.}
}
\value{
RData, CSV, and XLSX datafile with the desired variables & launches user interface.
}
\description{
Main function. Interacts with the user to select the desired variables for the report, loads
them, saves them in an external output, runs the verifications, and informs the user about the
success of the whole process. Either the `project_path` should be specified, or the `db_path`
with all the related items, being `db_name`, `prj_name`, and `scenarios`.
}
