% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_gen_map}
\alias{se_gen_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
se_gen_map
}
\description{
primary, secondary, final energy maps
\dontrun{
 library(gcamreport);
 gcamreport::se_gen_map
}
}
\keyword{datasets}
