% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector.list}
\alias{sector.list}
\title{sector.list}
\format{
An object of class \code{character} of length 8.
}
\usage{
sector.list
}
\description{
categorize sector
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::sector.list
}
}
\keyword{datasets}
