% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{soil_alloc_exp_func}
\alias{soil_alloc_exp_func}
\title{soil_alloc_exp_func}
\usage{
soil_alloc_exp_func(y, y_conv, E_soil_t, s)
}
\arguments{
\item{y}{Calendar year of emission accounting.}

\item{y_conv}{Year of land-use conversion.}

\item{s}{Soil turnover timescale parameter (years); half-life = s/10 * ln2.}

\item{E_veg_t}{Total potential soil carbon change (tCO2/ha).}
}
\value{
Annual soil carbon flux (tCO2/ha/yr) for year y.
}
\description{
Soil carbon: exponential allocation function ----
GCAM allocates soil carbon change using a first-order exponential decay
function parameterised by a soil timescale (s). The annual flux is the
incremental change in cumulative fraction between consecutive years.
}
\keyword{internal}
