% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{src_dir}
\alias{src_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
src_dir
}
\description{
source directory
\dontrun{
 library(gcamreport);
 gcamreport::src_dir
}
}
\keyword{datasets}
