% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_names}
\alias{state_names}
\title{description}
\format{
An object of class \code{character} of length 51.
}
\usage{
state_names
}
\description{
state names
\dontrun{
 library(gcamreport);
 gcamreport::state_names
}
}
\keyword{datasets}
