% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_crop_production.R
\name{summarize_crop_production}
\alias{summarize_crop_production}
\title{Get Crop Production Summary by Province}
\usage{
summarize_crop_production(provincial_prod, year_filter = 2020)
}
\arguments{
\item{provincial_prod}{Provincial production data}

\item{year_filter}{Optional year to filter (default: 2020)}
}
\value{
Summary table
}
\description{
Get Crop Production Summary by Province
}
