% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_livestock_production.R
\name{summarize_livestock_production}
\alias{summarize_livestock_production}
\title{Get Livestock Production Summary by Province}
\usage{
summarize_livestock_production(provincial_prod, year_filter = 2020)
}
\arguments{
\item{provincial_prod}{Provincial production data}

\item{year_filter}{Optional year to filter (default: 2020)}
}
\value{
Summary table
}
\description{
Get Livestock Production Summary by Province
}
