% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
template
}
\description{
read in template
\dontrun{
 library(gcamreport);
 gcamreport::template
}
}
\keyword{datasets}
