% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_v6.0}
\alias{template_v6.0}
\title{template_v6.0}
\format{
.csv
}
\source{
github
}
\usage{
template_v6.0
}
\description{
read in template compatible with GCAM6.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::template_v6.0
}
}
\keyword{datasets}
