% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
<<<<<<<< HEAD:man/ghg_GWP_AR5.Rd
\name{ghg_GWP_AR5}
\alias{ghg_GWP_AR5}
\title{ghg_GWP_AR5}
========
\name{template_v7.0}
\alias{template_v7.0}
\title{template_v7.0}
>>>>>>>> dev_newstructure:man/template_v7.0.Rd
\format{
.csv
}
\source{
local
}
\usage{
<<<<<<<< HEAD:man/ghg_GWP_AR5.Rd
ghg_GWP_AR5
}
\description{
ghg adjuster - uses 100-yr GWPs from AR5 in GCAM unit
========
template_v7.0
}
\description{
read in template compatible with GCAM7.0.
>>>>>>>> dev_newstructure:man/template_v7.0.Rd
}
\examples{
\dontrun{
library(gcamreport)
<<<<<<<< HEAD:man/ghg_GWP_AR5.Rd
gcamreport::ghg_GWP_AR5
========
gcamreport::template_v7.0
>>>>>>>> dev_newstructure:man/template_v7.0.Rd
}
}
\keyword{datasets}
