% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_vGCAMChina8.0}
\alias{template_vGCAMChina8.0}
\title{template_vGCAMChina8.0}
\format{
data frame
}
\usage{
template_vGCAMChina8.0
}
\description{
Template for GCAM-China 8.0
}
\keyword{datasets}
