% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_en_service}
\alias{transport_en_service}
\title{transport_en_service}
\format{
.csv
}
\source{
github
}
\usage{
transport_en_service
}
\description{
energy services
\dontrun{
 library(gcamreport);
 gcamreport::transport_en_service
}
}
\keyword{datasets}
