% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_en_service_v7.0}
\alias{transport_en_service_v7.0}
\title{transport_en_service_v7.0}
\format{
.csv
}
\source{
github
}
\usage{
transport_en_service_v7.0
}
\description{
transport energy services compatible with GCAM7.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::transport_en_service_v7.0
}
}
\keyword{datasets}
