% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_en_service_v7.1}
\alias{transport_en_service_v7.1}
\title{transport_en_service_v7.1}
\format{
.csv
}
\source{
github
}
\usage{
transport_en_service_v7.1
}
\description{
transport energy services compatible with GCAM7.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::transport_en_service_v7.1
}
}
\keyword{datasets}
