% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_final_en_map}
\alias{transport_final_en_map}
\title{transport_final_en_map}
\format{
.csv
}
\source{
github
}
\usage{
transport_final_en_map
}
\description{
primary, secondary, final energy maps
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::transport_final_en_map
}
}
\keyword{datasets}
