% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transport_final_en_map_v7.1}
\alias{transport_final_en_map_v7.1}
\title{transport_final_en_map_v7.1}
\format{
.csv
}
\source{
github
}
\usage{
transport_final_en_map_v7.1
}
\description{
primary, secondary, final energy maps compatible with GCAM7.1.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::transport_final_en_map_v7.1
}
}
\keyword{datasets}
