% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trn_truck_annual_travel}
\alias{trn_truck_annual_travel}
\title{trn_truck_annual_travel}
\format{
.csv
}
\source{
local assumption; inst/extdata/mappings/common/trn_truck_annual_travel.csv
}
\usage{
trn_truck_annual_travel
}
\description{
annual distance per truck (vkt/veh/yr) by GCAM truck size class, used for truck stocks and sales
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::trn_truck_annual_travel
}
}
\keyword{datasets}
