% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{update_user_choices_plot}
\alias{update_user_choices_plot}
\title{update_user_choices_plot}
\usage{
update_user_choices_plot(
  selected_scen,
  selected_years,
  tree_regions,
  tree_variables,
  sidebarItemExpanded
)
}
\arguments{
\item{selected_scen}{user's selected scenarios}

\item{selected_years}{user's selected years}

\item{tree_regions}{user's selected regions}

\item{tree_variables}{user's selected variables}

\item{sidebarItemExpanded}{input sidebar expanded item}
}
\description{
Update dataframe to display plots with the user's choices
}
\keyword{internal}
