% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcam2gains_buildings.R
\name{validate_gains_conversion}
\alias{validate_gains_conversion}
\title{Validate GCAM2GAINS Conversion}
\usage{
validate_gains_conversion(gains_data, gcam_data)
}
\arguments{
\item{gains_data}{Converted GAINS data}

\item{gcam_data}{Original GCAM data}
}
\value{
List with validation results
}
\description{
Validate GCAM2GAINS Conversion
}
