% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_fun_map}
\alias{var_fun_map}
\title{description}
\format{
.csv
}
\source{
local
}
\usage{
var_fun_map
}
\description{
mapping between variables, functions to load them, dependent variables, and availables verifications
\dontrun{
 library(report);
 report::var_fun_map
}
}
\keyword{datasets}
