% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{veg_alloc_S_curve}
\alias{veg_alloc_S_curve}
\title{veg_alloc_S_curve}
\usage{
veg_alloc_S_curve(y, y_conv, E_veg_t, M)
}
\arguments{
\item{y}{Calendar year of emission accounting.}

\item{y_conv}{Year of land-use conversion}

\item{E_veg_t}{Total potential soil carbon change (tCO2/ha)}

\item{M}{Mature age of vegetation (years); controls the growth rate}
}
\value{
Annual vegetation carbon flux (tCO2/ha/yr) for year y.
}
\description{
Vegetation carbon: sigmoid allocation function ----#
GCAM allocates vegetation carbon change using a sigmoid (S-curve) growth
function. The annual flux is the incremental change in cumulative fraction
between consecutive years.
}
\keyword{internal}
