% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vet_dir}
\alias{vet_dir}
\title{description}
\format{
An object of class \code{character} of length 1.
}
\usage{
vet_dir
}
\description{
vetting directory
\dontrun{
 library(report);
 report::vet_dir
}
}
\keyword{datasets}
