% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{approx_fun}
\alias{approx_fun}
\title{approx_fun}
\usage{
approx_fun(year, value, rule = 1)
}
\arguments{
\item{year}{Year(s) for interpolation or extrapolation.}

\item{value}{Values to interpolate or extrapolate from.}

\item{rule}{Number of points to use for extrapolation.}
}
\description{
Performs interpolation or extrapolation on given values.
}
\keyword{internal}
