% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{approx_fun}
\alias{approx_fun}
\title{approx_fun}
\usage{
approx_fun(year, value, rule = 1)
}
\arguments{
\item{year:}{year to consider}

\item{value:}{values to extrapolate from}

\item{rule:}{number of points to extrapolate}
}
\description{
Interpolation function
}
