% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcm_to_EJ}
\alias{bcm_to_EJ}
\title{description}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
bcm_to_EJ
}
\description{
units conversion
\dontrun{
 library(report);
 report::bcm_to_EJ
}
}
\keyword{datasets}
