% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon_content}
\alias{carbon_content}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
carbon_content
}
\description{
carbon content
\dontrun{
 library(gcamreport);
 gcamreport::carbon_content
}
}
\keyword{datasets}
