% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon_seq_tech_map}
\alias{carbon_seq_tech_map}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
carbon_seq_tech_map
}
\description{
emissions maps
\dontrun{
 library(report);
 report::carbon_seq_tech_map
}
}
\keyword{datasets}
