% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_75USD_10USD}
\alias{conv_75USD_10USD}
\title{description}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
conv_75USD_10USD
}
\description{
units conversion
\dontrun{
 library(report);
 report::conv_75USD_10USD
}
}
\keyword{datasets}
