% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{conv_EJ_GW}
\alias{conv_EJ_GW}
\title{conv_EJ_GW}
\usage{
conv_EJ_GW(data, cf, EJ)

conv_EJ_GW(data, cf, EJ)

conv_EJ_GW(data, cf, EJ)
}
\arguments{
\item{data}{dataset}

\item{cf}{conversion factor}

\item{EJ}{EJ amount}
}
\description{
Covert EJ to GW

Covert EJ to GW

Covert EJ to GW
}
\keyword{conversion}
\keyword{internal}
