% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{conv_ghg_co2e}
\alias{conv_ghg_co2e}
\title{conv_ghg_co2e}
\usage{
conv_ghg_co2e(data)

conv_ghg_co2e(data)

conv_ghg_co2e(data)
}
\arguments{
\item{data}{dataset}
}
\description{
Covert GHG to CO2e

Covert GHG to CO2e

Covert GHG to CO2e
}
\keyword{conversion}
\keyword{internal}
