% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_million_billion}
\alias{conv_million_billion}
\title{conv_million_billion}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
conv_million_billion
}
\description{
units conversion
\dontrun{
 library(gcamreport);
 gcamreport::conv_million_billion
}
}
\keyword{datasets}
