% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{convert}
\alias{convert}
\title{convert}
\format{
An object of class \code{list} of length 12.
}
\usage{
convert
}
\description{
units conversion list
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::convert
}
}
\keyword{datasets}
