% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(db_path, db_name, prj_name, scenarios)
}
\arguments{
\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}
}
\value{
loaded project into global environment
}
\description{
Create specified project and load it into the global environment
}
