% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(
  db_path,
  db_name,
  prj_name,
  scenarios,
  desired_regions = "All",
  desired_variables = "All"
)
}
\arguments{
<<<<<<< HEAD
\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}
=======
\item{db_path}{path of the database.}

\item{db_name}{name of the database.}

\item{prj_name}{name of the project.}

\item{scenarios}{name of the scenarios to be considered. By default, all the scenarios in the database are considered.}
>>>>>>> e5eba6c (Allow default selection of all db/prj scenarios. Improve user's input checks. Add tests)

\item{desired_regions:}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
In case the project dataset needs to be created, it will be produced with only the specified regions.}

\item{desired_variables:}{desired variables to have in the report. Considered 'All' by default.
Otherwise, specify a vector with all the desired options. To know all possible options, run `available_variables()`.
In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3".}
}
\value{
loaded project into global environment.
}
\description{
Create specified project and load it into the global environment
}
