% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_project}
\alias{create_project}
\title{create_project}
\usage{
create_project(
  db_path,
  db_name,
  prj_name,
  scenarios = NULL,
  desired_regions = "All",
  desired_variables = "All",
  GCAM_version = "v7.0",
  queries_general_file = NULL,
  queries_nonCO2_file = NULL
)
}
\arguments{
\item{db_path}{path of the database.}

\item{db_name}{name of the database.}

\item{prj_name}{name of the project.}

\item{scenarios}{name of the scenarios to be considered. By default, all the scenarios in the database are considered.}

\item{desired_regions}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
In case the project dataset needs to be created, it will be produced with only the specified regions.}

\item{desired_variables}{desired variables to have in the report. Considered 'All' by default.
Otherwise, specify a vector with all the desired options. To know all possible options, run `available_variables()`.
In case the project dataset needs to be created, it will be produced with only the specified variables. ATTENTION:
the global variables such as "Emissions" will be computed considering only the selected variables, for instance "Emissions|CO2",
and will no account for other variables, such as "Emissions|CH4" or "Emissions|NH3".}

\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}

\item{queries_general_file}{full path to a general xml query file (including file name and extension). By default it points to the
gcamreport general query file, compatible with the latest GCAM version and able to report all standardized variables.}

\item{queries_nonCO2_file}{full path to an xml query file (including file name and extension) designed to load long nonCO2 queries:
"nonCO2 emissions by sector (excluding resource production)" and "nonCO2 emissions by region". By default it points to the
gcamreport nonCO2 query file, compatible with the latest GCAM version and necessary to report some of the standardized variables.}
}
\value{
loaded project into global environment and local saved project.
}
\description{
Create specified project and load it into the global environment
}
