% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CUMULATIVE_EMISSIONS_BUDGET_GOALS}
\alias{CUMULATIVE_EMISSIONS_BUDGET_GOALS}
\title{CUMULATIVE_EMISSIONS_BUDGET_GOALS}
\format{
An object of class \code{numeric} of length 16.
}
\usage{
CUMULATIVE_EMISSIONS_BUDGET_GOALS
}
\description{
cumulative emissions budget goals list
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::CUMULATIVE_EMISSIONS_BUDGET_GOALS
}
}
\keyword{datasets}
