% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(
  type,
  db_path,
  db_name,
  prj_name,
  scenarios,
  desired_regions = "All",
  GCAM_version = "v7.0",
  queries_nonCO2_file
)
}
\arguments{
\item{type}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'.}

\item{db_path}{path of the database.}

\item{db_name}{name of the database.}

\item{prj_name}{name of the project.}

\item{scenarios}{name of the scenarios to be considered.}

\item{desired_regions}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
In case the project dataset needs to be created, it will be produced with only the specified regions.}

\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}

\item{queries_nonCO2_file}{full path to an xml query file (including file name and extension) designed to load long nonCO2 queries:
"nonCO2 emissions by sector (excluding resource production)" and "nonCO2 emissions by region". By default it points to the
gcamreport nonCO2 query file, compatible with the latest GCAM version and necessary to report some of the standardized variables.}
}
\value{
dataframe with the data from the query.
}
\description{
Add nonCO2 large queries
}
