% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(type, db_path, db_name, prj_name, scenarios)
}
\arguments{
\item{db_path:}{path of the database}

\item{db_name:}{name of the database}

\item{prj_name:}{name of the project}

\item{scenarios:}{name of the scenarios to be considered}

\item{type:}{either 'nonCO2 emissions by region' or 'nonCO2 emissions by sector'}
}
\value{
dataframe with the data from the queries
}
\description{
Add nonCO2 large queries
}
