% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_bind_results}
\alias{do_bind_results}
\title{do_bind_results}
\usage{
do_bind_results(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
Saved results in an output file.
}
\description{
Binds results and saves them to an output file.#'
}
\keyword{internal}
\keyword{process}
