% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_check_trade}
\alias{do_check_trade}
\title{do_check_trade}
\usage{
do_check_trade()
}
\value{
Verification message indicating if the process was successful.
}
\description{
Check global trade is zero.
}
\keyword{check}
