% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{do_check_trade}
\alias{do_check_trade}
\title{do_check_trade}
\usage{
do_check_trade()

do_check_trade()

do_check_trade()
}
\value{
Verification message indicating if the process was successful.

Verification message indicating if the process was successful.

Verification message indicating if the process was successful.
}
\description{
Check global trade is zero.

Check global trade is zero.

Check global trade is zero.
}
\keyword{check}
\keyword{internal}
