% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_check_vetting}
\alias{do_check_vetting}
\title{do_check_vetting}
\usage{
do_check_vetting(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
A message confirming the success of the vetting process.
}
\description{
Verify vetting and produce plot.
}
\keyword{check}
\keyword{internal}
