% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{do_check_vetting}
\alias{do_check_vetting}
\title{do_check_vetting}
\usage{
do_check_vetting()

do_check_vetting()

do_check_vetting()
}
\value{
Verification message indicating if the process was successful.

Verification message indicating if the process was successful.

Verification message indicating if the process was successful.
}
\description{
Verify vetting and produce plot.

Verify vetting and produce plot.

Verify vetting and produce plot.
}
\keyword{check}
\keyword{internal}
