% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EJ_to_GWh}
\alias{EJ_to_GWh}
\title{EJ_to_GWh}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EJ_to_GWh
}
\description{
units conversion
\dontrun{
 library(gcamreport);
 gcamreport::EJ_to_GWh
}
}
\keyword{datasets}
