% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fill_queries}
\alias{fill_queries}
\title{fill_queries}
\usage{
fill_queries()
}
\value{
empty dataframes on the void queries of the project
}
\description{
Create a folder to save the datasets and file, in case it does not exist
}
