% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fillScaleSector}
\alias{fillScaleSector}
\title{fillScaleSector}
\format{
An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.
}
\usage{
fillScaleSector
}
\description{
final filling color scale
\dontrun{
 library(gcamreport);
 gcamreport::fillScaleSector
}
}
\keyword{datasets}
