% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_data_regions}
\alias{filter_data_regions}
\title{filter_data_regions}
\usage{
filter_data_regions(data)
}
\value{
data containing only the desired regions
}
\description{
filter the desired regions of some data with "regions" column
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
