% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{filter_data_regions}
\alias{filter_data_regions}
\title{filter_data_regions}
\usage{
filter_data_regions(data)

filter_data_regions(data)

filter_data_regions(data)
}
\value{
data containing only the desired regions

data containing only the desired regions

data containing only the desired regions
}
\description{
Filter the desired regions of some data with "regions" column

Filter the desired regions of some data with "regions" column

Filter the desired regions of some data with "regions" column
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
