% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_loading_regions}
\alias{filter_loading_regions}
\title{filter_loading_regions}
\usage{
filter_loading_regions(data, desired_regions = "All", variable)
}
\arguments{
\item{data}{dataframe to be filtered}

\item{desired_regions}{desired regions to consider. By default, 'All'. Otherwise, specify a vector with all the considered regions.
To know all possible regions, run `available_regions()`. ATTENTION: the considered regions will make up "World".
In case the project dataset needs to be created, it will be produced with only the specified regions.}

\item{variable}{dataset variable information}
}
\value{
filtered dataframe
}
\description{
Filter the desired regions of a GCAM project
}
