% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel.list}
\alias{fuel.list}
\title{fuel.list}
\format{
An object of class \code{character} of length 8.
}
\usage{
fuel.list
}
\description{
fuels list
\dontrun{
 library(gcamreport);
 gcamreport::fuel.list
}
}
\keyword{datasets}
