% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gather_map}
\alias{gather_map}
\title{gather_map}
\usage{
gather_map(df)
}
\description{
Formats all maps into a long table
}
<<<<<<< HEAD
=======
\keyword{internal}
>>>>>>> f4b3e58 (Review: fix global data items warnings and clean unnecessary items)
