% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_years}
\alias{GCAM_years}
\title{GCAM_years}
\format{
An object of class \code{numeric} of length 21.
}
\usage{
GCAM_years
}
\description{
reporting years
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::GCAM_years
}
}
\keyword{datasets}
