% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GCAM_years}
\alias{GCAM_years}
\title{description}
\format{
An object of class \code{numeric} of length 21.
}
\usage{
GCAM_years
}
\description{
reporting years
\dontrun{
 library(report);
 report::GCAM_years
}
}
\keyword{datasets}
