% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_ag_demand}
\alias{get_ag_demand}
\title{get_ag_demand}
\usage{
get_ag_demand()
}
\value{
ag_demand_clean global variable
}
\description{
Get agricultural demand.
}
\keyword{ag}
\keyword{getQuery}
