% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ag_demand}
\alias{get_ag_demand}
\title{get_ag_demand}
\usage{
get_ag_demand(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`ag_demand_clean` global variable.
}
\description{
Get agricultural demand.
}
\keyword{ag}
\keyword{internal}
