% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ag_prices}
\alias{get_ag_prices}
\title{get_ag_prices}
\usage{
get_ag_prices(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
ag_prices_clean global variable
}
\description{
Calculate average mean for ag global index
}
\keyword{ag}
\keyword{internal}
