% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_ag_prices}
\alias{get_ag_prices}
\title{get_ag_prices}
\usage{
get_ag_prices()

get_ag_prices()

get_ag_prices()
}
\value{
ag_prices_clean global variable

ag_prices_clean global variable

ag_prices_clean global variable
}
\description{
Calculate average mean for ag global index

Calculate average mean for ag global index

Calculate average mean for ag global index
}
\keyword{ag}
\keyword{internal}
