% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_CCS_invest}
\alias{get_CCS_invest}
\title{get_CCS_invest}
\usage{
get_CCS_invest(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
CCS_invest_clean global variable
}
\description{
Calculate CSS investment
}
\keyword{internal}
\keyword{investment}
\keyword{process}
