% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_CCS_invest}
\alias{get_CCS_invest}
\title{get_CCS_invest}
\usage{
get_CCS_invest()

get_CCS_invest()

get_CCS_invest()
}
\value{
CCS_invest_clean global variable

CCS_invest_clean global variable

CCS_invest_clean global variable
}
\description{
Calculate CSS investment

Calculate CSS investment

Calculate CSS investment
}
\keyword{internal}
\keyword{investment}
\keyword{process}
