% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_cf_iea_tmp}
\alias{get_cf_iea_tmp}
\title{get_cf_iea_tmp}
\usage{
get_cf_iea_tmp()
}
\value{
cf_iea global variable
}
\description{
Calculate cf for existing capacity checking global existing capacity from IEA.
}
\keyword{capacity}
\keyword{process}
\keyword{tmp}
