% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_cf_iea_tmp}
\alias{get_cf_iea_tmp}
\title{get_cf_iea_tmp}
\usage{
get_cf_iea_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`cf_iea` global variable
}
\description{
Calculate cf for existing capacity checking global existing capacity from IEA.
}
\keyword{capacity}
\keyword{internal}
\keyword{process}
\keyword{tmp}
