% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2}
\alias{get_co2}
\title{get_co2}
\usage{
get_co2()
}
\value{
co2_clean global variable
}
\description{
Get World's CO2 emissions query.
}
\keyword{co2}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
