% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2}
\alias{get_co2}
\title{get_co2}
\usage{
get_co2()

get_co2()

get_co2()
}
\value{
co2_clean global variable

co2_clean global variable

co2_clean global variable
}
\description{
Get World's CO2 emissions query.

Get World's CO2 emissions query.

Get World's CO2 emissions query.
}
\keyword{co2}
\keyword{internal}
