% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_concentration}
\alias{get_co2_concentration}
\title{get_co2_concentration}
\usage{
get_co2_concentration(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version: 'v7.0' (default) or 'v6.0'.}
}
\value{
`co2_concentration_clean` global variable.
}
\description{
Retrieves the global CO2 concentration query.
}
\keyword{co2}
\keyword{internal}
