% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2_concentration}
\alias{get_co2_concentration}
\title{get_co2_concentration}
\usage{
get_co2_concentration()

get_co2_concentration()

get_co2_concentration()
}
\value{
co2_concentration_clean global variable

co2_concentration_clean global variable

co2_concentration_clean global variable
}
\description{
Get World's CO2 concentration query.

Get World's CO2 concentration query.

Get World's CO2 concentration query.
}
\keyword{co2}
\keyword{internal}
