% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_emissions}
\alias{get_co2_emissions}
\title{get_co2_emissions}
\usage{
get_co2_emissions()
}
\value{
co2_emissions_clean global variable
}
\description{
Combine CO2 emission queries.
}
\keyword{co2}
\keyword{process}
