% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_ets}
\alias{get_co2_ets}
\title{get_co2_ets}
\usage{
get_co2_ets()
}
\value{
co2_ets global variable
}
\description{
Get World's CO2 ETS emissions query.
}
\keyword{co2}
\keyword{rgcam::getQuery}
