% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_iron_steel}
\alias{get_co2_iron_steel}
\title{get_co2_iron_steel}
\usage{
get_co2_iron_steel(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version to use: v7.0 (default) or v6.0.}
}
\value{
`co2_tech_ironsteel` global variable.
}
\description{
Retrieves CO2 emissions data for iron and steel.
}
\keyword{co2}
\keyword{internal}
\keyword{iron}
\keyword{steel}
