% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price}
\alias{get_co2_price}
\title{get_co2_price}
\usage{
get_co2_price()
}
\value{
co2_price_clean global variable
}
\description{
Get co2 price.
}
\keyword{co2}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
