% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price}
\alias{get_co2_price}
\title{get_co2_price}
\usage{
get_co2_price(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
co2_price_clean global variable
}
\description{
Get co2 price.
}
\keyword{co2}
\keyword{internal}
