% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price_fragmented_tmp}
\alias{get_co2_price_fragmented_tmp}
\title{get_co2_price_fragmented_tmp}
\usage{
get_co2_price_fragmented_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
`co2_price_fragmented` global variable
}
\description{
Retrieve CO2 fragmented price.
}
\keyword{co2}
\keyword{internal}
\keyword{tmp}
