% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price_fragmented_tmp}
\alias{get_co2_price_fragmented_tmp}
\title{get_co2_price_fragmented_tmp}
\usage{
get_co2_price_fragmented_tmp()
}
\value{
co2_price_fragmented global variable
}
\description{
Get fragmented co2 price.
}
\keyword{co2}
\keyword{rgcam::getQuery}
\keyword{tmp}
