% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_price_global_tmp}
\alias{get_co2_price_global_tmp}
\title{get_co2_price_global_tmp}
\usage{
get_co2_price_global_tmp()
}
\value{
co2_price_global & regions global variable
}
\description{
Get global co2 price.
}
\keyword{co2}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
\keyword{tmp}
