% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2_price_global_tmp}
\alias{get_co2_price_global_tmp}
\title{get_co2_price_global_tmp}
\usage{
get_co2_price_global_tmp()

get_co2_price_global_tmp()

get_co2_price_global_tmp()
}
\value{
co2_price_global & regions global variable

co2_price_global & regions global variable

co2_price_global & regions global variable
}
\description{
Get global co2 price.

Get global co2 price.

Get global co2 price.
}
\keyword{co2}
\keyword{internal}
\keyword{tmp}
