% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_sequestration}
\alias{get_co2_sequestration}
\title{get_co2_sequestration}
\usage{
get_co2_sequestration(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
co2_sequestration_clean global variable
}
\description{
Get carbon sequestration.
}
\keyword{co2}
\keyword{internal}
