% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_sequestration}
\alias{get_co2_sequestration}
\title{get_co2_sequestration}
\usage{
get_co2_sequestration()
}
\value{
co2_sequestration_clean global variable
}
\description{
Get carbon sequestration.
}
\keyword{co2}
\keyword{getQuery}
