% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_co2_sequestration}
\alias{get_co2_sequestration}
\title{get_co2_sequestration}
\usage{
get_co2_sequestration()

get_co2_sequestration()

get_co2_sequestration()
}
\value{
co2_sequestration_clean global variable

co2_sequestration_clean global variable

co2_sequestration_clean global variable
}
\description{
Get carbon sequestration.

Get carbon sequestration.

Get carbon sequestration.
}
\keyword{co2}
\keyword{internal}
