% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_tech_emissions_tmp}
\alias{get_co2_tech_emissions_tmp}
\title{get_co2_tech_emissions_tmp}
\usage{
get_co2_tech_emissions_tmp()
}
\value{
co2_tech_emissions global variable
}
\description{
Get no bio CO2 emissions query by sector, subsector, and techonolgy.
}
\keyword{co2}
\keyword{tmp}
