% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_tech_emissions_tmp}
\alias{get_co2_tech_emissions_tmp}
\title{get_co2_tech_emissions_tmp}
\usage{
get_co2_tech_emissions_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version to use: 'v7.0' (default) or 'v6.0'.}
}
\value{
`co2_tech_emissions` global variable.
}
\description{
Retrieves the non-bio CO2 emissions query by sector, subsector, and technology.
}
\keyword{co2}
\keyword{internal}
\keyword{tmp}
