% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_co2_tech_nobio_tmp}
\alias{get_co2_tech_nobio_tmp}
\title{get_co2_tech_nobio_tmp}
\usage{
get_co2_tech_nobio_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`co2_tech_nobio` global variable.
}
\description{
Retrieves the non-bio CO2 emissions query by sector and technology.
}
\keyword{co2}
\keyword{internal}
\keyword{tmp}
