% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_capacity_add}
\alias{get_elec_capacity_add}
\title{get_elec_capacity_add}
\usage{
get_elec_capacity_add(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`elec_capacity_add_clean` global variable
}
\description{
Calculate final total added capacity.
}
\keyword{capacity}
\keyword{internal}
\keyword{process}
