% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_elec_capacity_tot}
\alias{get_elec_capacity_tot}
\title{get_elec_capacity_tot}
\usage{
get_elec_capacity_tot()
}
\value{
elec_capacity_tot_clean global variable
}
\description{
Calculate electricity total capacity
}
\keyword{capacity}
\keyword{process}
