% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_capacity_tot}
\alias{get_elec_capacity_tot}
\title{get_elec_capacity_tot}
\usage{
get_elec_capacity_tot(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
elec_capacity_tot_clean global variable
}
\description{
Calculate electricity total capacity
}
\keyword{capacity}
\keyword{internal}
\keyword{process}
