% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_capital}
\alias{get_elec_capital}
\title{get_elec_capital}
\usage{
get_elec_capital(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`elec_capital_clean``` global variable
}
\description{
Calculate electricity capital.
}
\keyword{capital}
\keyword{internal}
\keyword{process}
