% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_cf_tmp}
\alias{get_elec_cf_tmp}
\title{get_elec_cf_tmp}
\usage{
get_elec_cf_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
`elec_cf``` global variable
}
\description{
Computes future capacity estimates using GCAM.
}
\keyword{capacity}
\keyword{internal}
\keyword{process}
\keyword{tmp}
