% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_gen_tech}
\alias{get_elec_gen_tech}
\title{get_elec_gen_tech}
\usage{
get_elec_gen_tech()
}
\value{
elec_gen_tech_clean global variable
}
\description{
Get electricity generation
}
\keyword{electricity}
