% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_gen_tech}
\alias{get_elec_gen_tech}
\title{get_elec_gen_tech}
\usage{
get_elec_gen_tech(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`elec_gen_tech_clean` global variable
}
\description{
Retrieve electricity generation.
}
\keyword{electricity}
\keyword{internal}
