% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_elec_investment}
\alias{get_elec_investment}
\title{get_elec_investment}
\usage{
get_elec_investment()
}
\value{
elec_investment_clean global variable
}
\description{
Calculate electricity investment = annual capacity additions * capital costs
}
\keyword{capital}
\keyword{process}
