% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_price}
\alias{get_energy_price}
\title{get_energy_price}
\usage{
get_energy_price()
}
\value{
`energy_price_clean``` global variable
}
\description{
Compute final energy price
}
\keyword{internal}
\keyword{price}
\keyword{process}
