% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_price_fragmented}
\alias{get_energy_price_fragmented}
\title{get_energy_price_fragmented}
\usage{
get_energy_price_fragmented(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`energy_price_fragmented` global variable
}
\description{
Retrieves energy prices fragmented by region and joins them since prices vary by region.
}
\keyword{internal}
\keyword{prices}
\keyword{process}
