% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_energy_price_tmp}
\alias{get_energy_price_tmp}
\title{get_energy_price_tmp}
\usage{
get_energy_price_tmp()

get_energy_price_tmp()

get_energy_price_tmp()
}
\value{
energy_price global variable

energy_price global variable

energy_price global variable
}
\description{
Bind regional oil, gas, coal prices with other energy prices

Bind regional oil, gas, coal prices with other energy prices

Bind regional oil, gas, coal prices with other energy prices
}
\keyword{internal}
\keyword{price}
\keyword{tmp}
