% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_service_buildings}
\alias{get_energy_service_buildings}
\title{get_energy_service_buildings}
\usage{
get_energy_service_buildings(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
energy_service_buildings_clean global variable
}
\description{
Get ES buildings.
}
\keyword{energy}
\keyword{internal}
