% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_service_transportation}
\alias{get_energy_service_transportation}
\title{get_energy_service_transportation}
\usage{
get_energy_service_transportation(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
energy_service_transportation_clean global variable
}
\description{
Get transport.
}
\keyword{energy}
\keyword{internal}
