% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_energy_service_transportation}
\alias{get_energy_service_transportation}
\title{get_energy_service_transportation}
\usage{
get_energy_service_transportation()
}
\value{
energy_service_transportation_clean global variable
}
\description{
Get transport.
}
\keyword{energy}
\keyword{getQuery}
