% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_energy_service_transportation}
\alias{get_energy_service_transportation}
\title{get_energy_service_transportation}
\usage{
get_energy_service_transportation()

get_energy_service_transportation()

get_energy_service_transportation()
}
\value{
energy_service_transportation_clean global variable

energy_service_transportation_clean global variable

energy_service_transportation_clean global variable
}
\description{
Get transport.

Get transport.

Get transport.
}
\keyword{energy}
\keyword{internal}
