% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_energy_trade_prod}
\alias{get_energy_trade_prod}
\title{get_energy_trade_prod}
\usage{
get_energy_trade_prod()

get_energy_trade_prod()

get_energy_trade_prod()
}
\value{
energy_trade_prod global variable

energy_trade_prod global variable

energy_trade_prod global variable
}
\description{
Get energy trade.

Get energy trade.

Get energy trade.
}
\keyword{energy}
\keyword{internal}
