% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_energy_trade_tmp}
\alias{get_energy_trade_tmp}
\title{get_energy_trade_tmp}
\usage{
get_energy_trade_tmp()
}
\value{
`energy_trade_supply` global variable.
}
\description{
Retrieve energy trade supply data for calculating other variables.
}
\keyword{energy}
\keyword{internal}
\keyword{tmp}
