% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_fe_sector}
\alias{get_fe_sector}
\title{get_fe_sector}
\usage{
get_fe_sector()
}
\value{
fe_sector_clean global variable
}
\description{
Compute final energy.
GPK 1/6/2022 - because some reporting categories are mapped from both the sector-level and subsector-level queries,
we need a step of aggregation here. For example, international and domestic air are both assigned to
aviation; international is mapped from the sector-level query, and domestic is mapped from the subsector (mode).
Without this step there would be duplicate entries with different data for the same reporting categories.
}
\keyword{energy}
\keyword{process}
