% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fe_sector_tmp}
\alias{get_fe_sector_tmp}
\title{get_fe_sector_tmp}
\usage{
get_fe_sector_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
fe_sector global variable
}
\description{
Get final energy demand by sector
}
\keyword{energy}
\keyword{internal}
\keyword{tmp}
