% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fe_transportation_tmp}
\alias{get_fe_transportation_tmp}
\title{get_fe_transportation_tmp}
\usage{
get_fe_transportation_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`fe_transportation` global variable
}
\description{
Retrieve mode-specific transport final energy, including rail, ship, and domestic air.
}
\keyword{energy}
\keyword{internal}
\keyword{tmp}
