% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fe_transportation_tmp}
\alias{get_fe_transportation_tmp}
\title{get_fe_transportation_tmp}
\usage{
get_fe_transportation_tmp()
}
\value{
fe_transportation global variable
}
\description{
Get mode-specific transport final energy to break out rail, ship, and domestic air.
}
\keyword{energy}
\keyword{rgcam::getQuery}
\keyword{tmp}
