% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_fgas}
\alias{get_fgas}
\title{get_fgas}
\usage{
get_fgas()

get_fgas()

get_fgas()
}
\value{
f_gas_clean global variable

f_gas_clean global variable

f_gas_clean global variable
}
\description{
Compute F-Gases emissions.

Compute F-Gases emissions.

Compute F-Gases emissions.
}
\keyword{f-gases}
\keyword{internal}
\keyword{process}
