% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_forcing}
\alias{get_forcing}
\title{get_forcing}
\usage{
get_forcing(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version: 'v7.0' (default) or 'v6.0'.}
}
\value{
`forcing_clean` global variable.
}
\description{
Retrieves the global forcing query.
}
\keyword{forcing}
\keyword{internal}
