% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_krw}
\alias{get_gdp_krw}
\title{get_gdp_krw}
\usage{
get_gdp_krw()
}
\value{
mer_krw global variable
}
\description{
Get GDP (PPP) query, compute regional GDP, and change units to [10USD].
}
\keyword{GDP}
\keyword{internal}
