% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_정미.R
\name{get_gdp_krw}
\alias{get_gdp_krw}
\title{get_gdp_krw}
\usage{
get_gdp_krw()

get_gdp_krw()
}
\value{
mer_krw global variable

mer_krw global variable
}
\description{
Get GDP (MER) query, compute regional GDP, and change units to [15KRW].

Get GDP (MER) query, compute regional GDP, and change units to [15KRW].
}
\keyword{GDP}
\keyword{internal}
