% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_mer}
\alias{get_gdp_mer}
\title{get_gdp_mer}
\usage{
get_gdp_mer(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`GDP_MER_clean` global variable.
}
\description{
Retrieves GDP (MER) data and converts units to 10 USD.
}
\keyword{GDP}
\keyword{internal}
