% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_mer}
\alias{get_gdp_mer}
\title{get_gdp_mer}
\usage{
get_gdp_mer(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
GDP_MER_clean global variable
}
\description{
Get GDP (MER) query and change units to [10USD].
}
\keyword{GDP}
\keyword{internal}
