% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_gdp_mer}
\alias{get_gdp_mer}
\title{get_gdp_mer}
\usage{
get_gdp_mer()
}
\value{
GDP_MER_clean global variable
}
\description{
Get GDP (MER) query and change units to [10USD].
}
\keyword{GDP}
\keyword{getQuery}
