% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_ppp}
\alias{get_gdp_ppp}
\title{get_gdp_ppp}
\usage{
get_gdp_ppp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
GDP_PPP_clean global variable
}
\description{
Get GDP (PPP) query, compute regional GDP, and change units to [10USD].
}
\keyword{GDP}
\keyword{internal}
