% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gdp_ppp}
\alias{get_gdp_ppp}
\title{get_gdp_ppp}
\usage{
get_gdp_ppp()
}
\value{
GDP_PPP_clean global variable
}
\description{
Get GDP (PPP) query, compute regional GDP, and change units to [10USD].
}
\keyword{GDP}
\keyword{internal}
