% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ghg}
\alias{get_ghg}
\title{get_ghg}
\usage{
get_ghg(GCAM_version = "v7.0", GWP_version = "AR5")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}

\item{GWP_version}{Global Warming Potential values version: AR5 (default), AR6, or AR4.}
}
\value{
ghg_clean global variable
}
\description{
Get total GHG emissions.
}
\keyword{ghg}
\keyword{internal}
