% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ghg_sector}
\alias{get_ghg_sector}
\title{get_ghg_sector}
\usage{
get_ghg_sector()
}
\value{
ghg_sector_clean global variable
}
\description{
Get sectorial GHG emissions.
}
\keyword{ghg}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
