% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_ghg_sector}
\alias{get_ghg_sector}
\title{get_ghg_sector}
\usage{
get_ghg_sector()

get_ghg_sector()

get_ghg_sector()
}
\value{
ghg_sector_clean global variable

ghg_sector_clean global variable

ghg_sector_clean global variable
}
\description{
Get sectorial GHG emissions.

Get sectorial GHG emissions.

Get sectorial GHG emissions.
}
\keyword{ghg}
\keyword{internal}
