% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_industry_production}
\alias{get_industry_production}
\title{get_industry_production}
\usage{
get_industry_production()
}
\value{
industry_production_clean global variable
}
\description{
Get industry production.
}
\keyword{energy}
\keyword{internal}
