% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_exports}
\alias{get_iron_steel_exports}
\title{get_iron_steel_exports}
\usage{
get_iron_steel_exports(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
iron_steel_exports global variable
}
\description{
Get iron steel production.
}
\keyword{industry}
\keyword{internal}
\keyword{tmp}
