% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_imports}
\alias{get_iron_steel_imports}
\title{get_iron_steel_imports}
\usage{
get_iron_steel_imports(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
`iron_steel_imports` global variable
}
\description{
Retrieve iron steel imports.
}
\keyword{industry}
\keyword{internal}
\keyword{tmp}
