% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_iron_steel_map}
\alias{get_iron_steel_map}
\title{get_iron_steel_map}
\usage{
get_iron_steel_map()
}
\value{
iron_steel_map global variable
}
\description{
Get iron and steel emissions.
}
\keyword{getQuery}
\keyword{iron}
\keyword{steel}
