% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_iron_steel_map}
\alias{get_iron_steel_map}
\title{get_iron_steel_map}
\usage{
get_iron_steel_map()

get_iron_steel_map()

get_iron_steel_map()
}
\value{
iron_steel_map global variable

iron_steel_map global variable

iron_steel_map global variable
}
\description{
Get iron and steel emissions.

Get iron and steel emissions.

Get iron and steel emissions.
}
\keyword{internal}
\keyword{iron}
\keyword{steel}
