% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lu_co2}
\alias{get_lu_co2}
\title{get_lu_co2}
\usage{
get_lu_co2(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version to use: v7.0 (default) or v6.0.}
}
\value{
`LU_carbon_clean` global variable.
}
\description{
Retrieves land use CO2 emissions data.
}
\keyword{co2}
\keyword{internal}
\keyword{lu}
