% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nonbio_tmp}
\alias{get_nonbio_tmp}
\title{get_nonbio_tmp}
\usage{
get_nonbio_tmp(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version: 'v7.0' (default) or 'v6.0'.}
}
\value{
`nonbio_share` global variable.
}
\description{
Retrieves the non-bio CO2 emissions query by sector.
}
\keyword{co2}
\keyword{internal}
