% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nonbio_tmp}
\alias{get_nonbio_tmp}
\title{get_nonbio_tmp}
\usage{
get_nonbio_tmp()
}
\value{
nonbio_share global variable
}
\description{
Get no bio CO2 emissions query by sector.
}
\keyword{co2}
\keyword{getQuery}
