% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_nonbio_tmp}
\alias{get_nonbio_tmp}
\title{get_nonbio_tmp}
\usage{
get_nonbio_tmp()

get_nonbio_tmp()

get_nonbio_tmp()
}
\value{
nonbio_share global variable

nonbio_share global variable

nonbio_share global variable
}
\description{
Get no bio CO2 emissions query by sector.

Get no bio CO2 emissions query by sector.

Get no bio CO2 emissions query by sector.
}
\keyword{co2}
\keyword{internal}
