% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nonco2_emissions}
\alias{get_nonco2_emissions}
\title{get_nonco2_emissions}
\usage{
get_nonco2_emissions()
}
\value{
nonco2_clean global variable
}
\description{
Get non CO2 emissions query.
}
\keyword{internal}
\keyword{nonco2}
\keyword{rgcam::getQuery}
