% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_population}
\alias{get_population}
\title{get_population}
\usage{
get_population()
}
\value{
population_clean global variable
}
\description{
Get population query and change units to [million].
}
\keyword{population}
\keyword{rgcam::getQuery}
