% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_population}
\alias{get_population}
\title{get_population}
\usage{
get_population(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
population_clean global variable
}
\description{
Get population query and change units to [million].
}
\keyword{internal}
\keyword{population}
