% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_price_var_tmp}
\alias{get_price_var_tmp}
\title{get_price_var_tmp}
\usage{
get_price_var_tmp()
}
\value{
price_var global variable
}
\description{
Get price variables to compute carbon price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
