% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prices_subsector}
\alias{get_prices_subsector}
\title{get_prices_subsector}
\usage{
get_prices_subsector(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
prices_subsector global variable
}
\description{
Get energy prices - primary, secondary, and final
}
\keyword{internal}
\keyword{prices}
