% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prices_subsector}
\alias{get_prices_subsector}
\title{get_prices_subsector}
\usage{
get_prices_subsector(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`prices_subsector_pre` and `prices_subsector_biomass` global variables
}
\description{
Retreive primary, secondary, and final energy prices.
}
\keyword{internal}
\keyword{prices}
