% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_primary_energy}
\alias{get_primary_energy}
\title{get_primary_energy}
\usage{
get_primary_energy(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{GCAM version to use: 'v7.0' (default) or 'v6.0'.}
}
\value{
`primary_energy_clean` global variable.
}
\description{
Retrieve primary energy consumption data by technology.
}
\keyword{energy}
\keyword{internal}
