% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_primary_energy}
\alias{get_primary_energy}
\title{get_primary_energy}
\usage{
get_primary_energy()

get_primary_energy()

get_primary_energy()
}
\value{
primary_energy_clean global variable

primary_energy_clean global variable

primary_energy_clean global variable
}
\description{
Get primary energy consumption by tech.

Get primary energy consumption by tech.

Get primary energy consumption by tech.
}
\keyword{energy}
\keyword{internal}
