% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regional_emission}
\alias{get_regional_emission}
\title{get_regional_emission}
\usage{
get_regional_emission(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
regional_emission global variable
}
\description{
Compute regional nonCO2 emission: regional production * nonCO2 coef.
}
\keyword{internal}
\keyword{nonco2}
