% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regional_emission}
\alias{get_regional_emission}
\title{get_regional_emission}
\usage{
get_regional_emission()
}
\value{
regional_emission global variable
}
\description{
Compute regional nonCO2 emission: regional production * nonCO2 coef.
}
\keyword{internal}
\keyword{nonco2}
\keyword{rgcam::getQuery}
