% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regions_en_weight_tmp}
\alias{get_regions_en_weight_tmp}
\title{get_regions_en_weight_tmp}
\usage{
get_regions_en_weight_tmp()
}
\value{
`region_fe_weight`,`region_pe_weight`,`region_se_weight` global variables
}
\description{
Retrieve regions final/primary/energy energy weights to compute energy price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
