% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_regions_tmp}
\alias{get_regions_tmp}
\title{get_regions_tmp}
\usage{
get_regions_tmp()
}
\value{
regions global variable
}
\description{
Get regions to compute carbon price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
