% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_regions_tmp}
\alias{get_regions_tmp}
\title{get_regions_tmp}
\usage{
get_regions_tmp()

get_regions_tmp()

get_regions_tmp()
}
\value{
regions global variable

regions global variable

regions global variable
}
\description{
Get regions to compute carbon price.

Get regions to compute carbon price.

Get regions to compute carbon price.
}
\keyword{internal}
\keyword{process}
\keyword{tmp}
