% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_resource_investment}
\alias{get_resource_investment}
\title{get_resource_investment}
\usage{
get_resource_investment(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{compatible GCAM version: v7.0 (default) or v6.0.}
}
\value{
resource_investment_clean global variable
}
\description{
Calculate investment of resource production
}
\keyword{internal}
\keyword{investment}
\keyword{process}
