% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_secondary_solids}
\alias{get_secondary_solids}
\title{get_secondary_solids}
\usage{
get_secondary_solids()

get_secondary_solids()

get_secondary_solids()
}
\value{
secondary_solids global variable

secondary_solids global variable

secondary_solids global variable
}
\description{
Get secondary solids

Get secondary solids

Get secondary solids
}
\keyword{energy}
\keyword{internal}
