% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_temperature}
\alias{get_temperature}
\title{get_temperature}
\usage{
get_temperature(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{main GCAM compatible version: v7.0 (default) or v6.0.}
}
\value{
global_temp_clean global variable
}
\description{
Get World's mean temperature query.
}
\keyword{internal}
\keyword{temperature}
