% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_temperature}
\alias{get_temperature}
\title{get_temperature}
\usage{
get_temperature()
}
\value{
global_temp_clean global variable
}
\description{
Get World's mean temperature query.
}
<<<<<<< HEAD
\keyword{rgcam::getQuery}
=======
\keyword{internal}
>>>>>>> dc7f961 (Review: fix warning message undefined data)
\keyword{temperature}
