% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_total_revenue}
\alias{get_total_revenue}
\title{get_total_revenue}
\usage{
get_total_revenue()
}
\value{
total_revenue global variable
}
\description{
Compute total revenue: total production * global price.
}
\keyword{getQuery}
\keyword{revenue}
