% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_total_revenue}
\alias{get_total_revenue}
\title{get_total_revenue}
\usage{
get_total_revenue()

get_total_revenue()

get_total_revenue()
}
\value{
total_revenue global variable

total_revenue global variable

total_revenue global variable
}
\description{
Compute total revenue: total production * global price.

Compute total revenue: total production * global price.

Compute total revenue: total production * global price.
}
\keyword{internal}
\keyword{revenue}
