% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_transmission_invest}
\alias{get_transmission_invest}
\title{get_transmission_invest}
\usage{
get_transmission_invest(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{Main GCAM compatible version: 'v7.0' (default), 'v7.1', or 'v6.0'.}
}
\value{
`transmission_invest_clean``` global variable
}
\description{
Calculate Investment in Electricity Transmission and Distribution. Scales 2020 numbers based on the average of other model results from Mcollion et al. 2018.
Converts 2015 values to 2010 dollars.
}
\keyword{internal}
\keyword{investment}
\keyword{process}
