% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_diff.R,
%   R/functions_정미.R
\name{get_transmission_invest}
\alias{get_transmission_invest}
\title{get_transmission_invest}
\usage{
get_transmission_invest()

get_transmission_invest()

get_transmission_invest()
}
\value{
transmission_invest_clean global variable

transmission_invest_clean global variable

transmission_invest_clean global variable
}
\description{
Calculate investment electricity transmission and distribution
scale 2020 number - average of other model results from Mcollion et al. 2018
Need to convert 2015 to 2010 $

Calculate investment electricity transmission and distribution
scale 2020 number - average of other model results from Mcollion et al. 2018
Need to convert 2015 to 2010 $

Calculate investment electricity transmission and distribution
scale 2020 number - average of other model results from Mcollion et al. 2018
Need to convert 2015 to 2010 $
}
\keyword{internal}
\keyword{investment}
\keyword{process}
