% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GHG_gases}
\alias{GHG_gases}
\title{description}
\format{
An object of class \code{character} of length 17.
}
\usage{
GHG_gases
}
\description{
ghg emission conversion
\dontrun{
 library(gcamreport);
 gcamreport::GHG_gases
}
}
\keyword{datasets}
