% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWP_adjuster_v7.0}
\alias{GWP_adjuster_v7.0}
\title{GWP_adjuster_v7.0}
\format{
.csv
}
\source{
local
}
\usage{
GWP_adjuster_v7.0
}
\description{
ghg adjuster compatible with GCAM7.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::GWP_adjuster_v7.0
}
}
\keyword{datasets}
