% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{launch_gcamreport_ui}
\alias{launch_gcamreport_ui}
\title{launch_gcamreport_ui}
\usage{
launch_gcamreport_ui(data_path = NULL, data = NULL)
}
\arguments{
\item{data_path:}{RData dataset path containing the standardized data. You can obtain
this dataset by using the function `gcamreport::generate_report`.}

\item{data:}{dataset containing the standardized data. You can obtain
this dataset by using the function `gcamreport::generate_report`.}
}
\value{
launch shiny interactive ui
}
\description{
Launch shiny interactive user interface.
}
