% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_variable}
\alias{load_variable}
\title{load_variable}
\usage{
load_variable(var)
}
\arguments{
\item{var:}{variable to be loaded}
}
\value{
load variable
}
\description{
Recursive function to load the desired variable and its dependent variables
}
\keyword{internal}
