% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonCO2_content}
\alias{nonCO2_content}
\title{description}
\format{
.csv
}
\source{
github
}
\usage{
nonCO2_content
}
\description{
non CO2 content
\dontrun{
 library(gcamreport);
 gcamreport::nonCO2_content
}
}
\keyword{datasets}
