% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonCO2_content_v7.0}
\alias{nonCO2_content_v7.0}
\title{nonCO2_content_v7.0}
\format{
.csv
}
\source{
github
}
\usage{
nonCO2_content_v7.0
}
\description{
non CO2 content compatible with GCAM7.0.
}
\examples{
\dontrun{
library(gcamreport)
gcamreport::nonCO2_content_v7.0
}
}
\keyword{datasets}
