% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco2_emis_resource_map}
\alias{nonco2_emis_resource_map}
\title{nonco2_emis_resource_map}
\format{
.csv
}
\source{
github
}
\usage{
nonco2_emis_resource_map
}
\description{
emissions maps
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::nonco2_emis_resource_map
}
}
\keyword{datasets}
