% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_energy_map}
\alias{primary_energy_map}
\title{primary_energy_map}
\format{
.csv
}
\source{
github
}
\usage{
primary_energy_map
}
\description{
primary, secondary, final energy maps
\dontrun{
 library(gcamreport);
 gcamreport::primary_energy_map
}
}
\keyword{datasets}
