% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_queries}
\alias{read_queries}
\title{read_queries}
\usage{
read_queries(
  project_name = "gas_fin_updated",
  final_db_year = 2100,
  desired_variables = "All",
  save = TRUE,
  file_name = "final_data",
  launch_app = TRUE
)
}
\arguments{
\item{project_name:}{name of the project. By default = gas_fin_updated.}

\item{final_db_year:}{final year of the database. By default = 2100.}

\item{desired_variables:}{desired variables to have in the report. Considered 'All' by default.
Otherwise, specify a vector with all the desired options, being population_clean, GDP_MER_clean, GDP_PPP_clean,
global_temp_clean, forcing_clean, co2_concentration_clean, co2_emissions_clean, tot_co2_clean, co2_sequestration_clean,
ag_demand_clean, land_clean, primary_energy_clean, energy_trade_clean, elec_gen_tech_clean, elec_capacity_tot_clean,
elec_capacity_add_clean, se_gen_tech_clean, fe_sector_clean, energy_service_transportation_clean, energy_service_buildings_clean,
ag_prices_clean, industry_production_clean, elec_capital_clean, elec_investment_clean, transmission_invest_clean,
CCS_invest_clean, resource_investment_clean, nonco2_clean, co2_price_clean.}

\item{save:}{if TRUE, save reporting data. Do not save otherwise.}

\item{file_name:}{file name of the saved data. Not used if data not saved. Data saved in '/output/datasets/'}

\item{launch_app:}{if TRUE, open app. Do not launch app otherwise.}
}
\value{
saved? datafile with the desired variables & launched? user interface
}
\description{
Main function. Interacts with the user to select the desired variables for the report,
loads them, saves them in an external output, runs the verifications, and informs the
user about the success of the whole process.
}
\keyword{internal}
