% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sector.color}
\alias{sector.color}
\title{sector.color}
\format{
An object of class \code{character} of length 8.
}
\usage{
sector.color
}
\description{
color by fuel
}
\examples{
\dontrun{
 library(gcamreport);
 gcamreport::sector.color
}
}
\keyword{datasets}
