% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_with_pattern}
\alias{start_with_pattern}
\title{start_with_pattern}
\usage{
start_with_pattern(vector, pattern)
}
\arguments{
\item{vector:}{vector to check}

\item{pattern:}{pattern to consider}
}
\value{
subvector of `vector` whose elements start with `pattern`
}
\description{
Return the vector elements starting with the specified parameters
}
\keyword{internal}
\keyword{utils}
