% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tech.color}
\alias{tech.color}
\title{tech.color}
\format{
An object of class \code{character} of length 20.
}
\usage{
tech.color
}
\description{
techology color
\dontrun{
 library(gcamreport);
 gcamreport::tech.color
}
}
\keyword{datasets}
