% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tech.list}
\alias{tech.list}
\title{description}
\format{
An object of class \code{character} of length 20.
}
\usage{
tech.list
}
\description{
techonolgies list
\dontrun{
 library(report);
 report::tech.list
}
}
\keyword{datasets}
