% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_template}
\alias{update_template}
\title{update_template}
\usage{
update_template(GCAM_version = "v7.0")
}
\arguments{
\item{GCAM_version}{The GCAM version to use: 'v7.0' (default) or 'v6.0'.}
}
\value{
Updated template saved as both .rda and .csv files in the `inst/extdata` folder.
}
\description{
Update the template file by incorporating new reported variables and removing unreported ones.
}
\keyword{internal}
\keyword{template}
